% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{add_class}
\alias{add_class}
\title{Add class}
\usage{
add_class(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\description{
add_class() is an Add Class generic that adds information about a class.NA
}
