% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_add_class.R
\name{add_class.ready4_class_pt_lup}
\alias{add_class.ready4_class_pt_lup}
\title{Add class method applied to Class Prototype Lookup Table readyforwhatsnext S3 class.}
\usage{
\method{add_class}{ready4_class_pt_lup}(
  x,
  tb_row_idx,
  make_tb,
  dev_pckg_namespace,
  name_prefix,
  output_dir,
  file_exists_logic,
  ignore_ns_chr = NA_character_,
  required_pckg_chr_vec = NA_character_,
  class_in_cache_logic_chr = "stop"
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{tb_row_idx}{PARAM_DESCRIPTION}

\item{make_tb}{Make (a tibble)}

\item{dev_pckg_namespace}{PARAM_DESCRIPTION}

\item{name_prefix}{PARAM_DESCRIPTION}

\item{output_dir}{PARAM_DESCRIPTION}

\item{file_exists_logic}{PARAM_DESCRIPTION}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1), Default: 'NA'}

\item{required_pckg_chr_vec}{Required pckg (a character vector), Default: 'NA'}

\item{class_in_cache_logic_chr}{Class in cache logic (a character vector of length 1), Default: 'stop'}
}
\value{
Inst of ready4 class prototype (a lookup table)
}
\description{
add_class.ready4_class_pt_lup() is an Add Class method that adds information about a class. This method is implemented for the Class Prototype Lookup Table readyforwhatsnext S3 class.The function returns inst of ready4 class prototype (a lookup table).
}
\keyword{internal}
