% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author.ready4class_manifest}
\alias{author.ready4class_manifest}
\alias{author,ready4class_manifest-method}
\title{Author method applied to ready4 S3 class Manifest..}
\usage{
author.ready4class_manifest(
  x,
  dv_url_pfx_1L_chr = NULL,
  init_class_pt_lup = NULL,
  key_1L_chr = NULL,
  list_generics_1L_lgl = F,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)

\S4method{author}{ready4class_manifest}(
  x,
  dv_url_pfx_1L_chr = NULL,
  init_class_pt_lup = NULL,
  key_1L_chr = NULL,
  list_generics_1L_lgl = F,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{x}{An instance of ready4 S3 class Manifest.}

\item{dv_url_pfx_1L_chr}{Dataverse url prefix (a character vector of length one), Default: NULL}

\item{init_class_pt_lup}{Initial class prototype (a lookup table), Default: NULL}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{list_generics_1L_lgl}{List generics (a logical vector of length one), Default: F}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector), Default: 'NA'}

\item{req_pkgs_chr}{Require packages (a character vector), Default: 'NA'}

\item{self_serve_1L_lgl}{Self serve (a logical vector of length one), Default: F}

\item{self_serve_fn_ls}{Self serve (a list of functions), Default: NULL}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
X (ready4 S4 class for package metadata required for package set-up.)
}
\description{
author.ready4class_manifest() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 S3 class Manifest. The function is called for its side effects and does not return a value.
}
