% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorClasses.R
\name{authorClasses.ready4class_constructor}
\alias{authorClasses.ready4class_constructor}
\alias{authorClasses,ready4class_constructor-method}
\title{AuthorClasses method applied to ready4 S3 class Constructor Table..}
\usage{
authorClasses.ready4class_constructor(
  x,
  name_pfx_1L_chr,
  output_dir_1L_chr,
  file_exists_cdn_1L_chr = NULL,
  prototype_lup = NULL,
  dev_pkg_ns_1L_chr = ready4fun::get_dev_pkg_nm(),
  nss_to_ignore_chr,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup,
  object_type_lup,
  consent_1L_chr = NULL
)

\S4method{authorClasses}{ready4class_constructor}(
  x,
  name_pfx_1L_chr,
  output_dir_1L_chr,
  file_exists_cdn_1L_chr = NULL,
  prototype_lup = NULL,
  dev_pkg_ns_1L_chr = ready4fun::get_dev_pkg_nm(),
  nss_to_ignore_chr,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup,
  object_type_lup,
  consent_1L_chr = NULL
)
}
\arguments{
\item{x}{An instance of ready4 S3 class Constructor Table.}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one)}

\item{file_exists_cdn_1L_chr}{File exists condition (a character vector of length one), Default: NULL}

\item{prototype_lup}{Prototype (a lookup table), Default: NULL}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one), Default: ready4fun::get_dev_pkg_nm()}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector)}

\item{req_pkgs_chr}{Require packages (a character vector), Default: 'NA'}

\item{class_in_cache_cdn_1L_chr}{Class in cache condition (a character vector of length one), Default: 'stop'}

\item{abbreviations_lup}{Abbreviations (a lookup table)}

\item{object_type_lup}{Object type (a lookup table)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: NULL}
}
\description{
authorClasses.ready4class_constructor() is an AuthorClasses method that authors and saves files necessary for creating and documenting classes. This method is implemented for the ready4 S3 class Constructor Table. The function is called for its side effects and does not return a value.
}
