% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_create.R
\name{create_accessors}
\alias{create_accessors}
\title{Create accessors}
\usage{
create_accessors(
  slot_name_chr,
  set_only,
  parent,
  class_name,
  print_accessors,
  output_folder,
  ignore_ns_chr,
  required_pckg_chr_vec
)
}
\arguments{
\item{slot_name_chr}{Slot name (a character vector of length 1)}

\item{set_only}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION}

\item{class_name}{PARAM_DESCRIPTION}

\item{print_accessors}{PARAM_DESCRIPTION}

\item{output_folder}{PARAM_DESCRIPTION}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1)}

\item{required_pckg_chr_vec}{Required pckg (a character vector)}
}
\description{
create_accessors() is a Create function that creates a new R object. Specifically, this function implements an algorithm to create accessors.NA
}
\keyword{internal}
