% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_create.R
\name{create_ready_accessors}
\alias{create_ready_accessors}
\title{Create ready accessors}
\usage{
create_ready_accessors(
  class_name,
  parent,
  print_accessors,
  output_folder,
  ignore_ns_chr,
  required_pckg_chr_vec,
  parent_ns_ls
)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION}

\item{print_accessors}{PARAM_DESCRIPTION}

\item{output_folder}{PARAM_DESCRIPTION}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1)}

\item{required_pckg_chr_vec}{Required pckg (a character vector)}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
create_ready_accessors() is a Create function that creates a new R object. Specifically, this function implements an algorithm to create ready accessors. The function is called for its side effects and does not return a value.
}
\keyword{internal}
