% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_create.R
\name{create_ready_helper}
\alias{create_ready_helper}
\title{Create ready helper}
\usage{
create_ready_helper(
  class_name,
  parent,
  class_slots,
  proto_ls,
  prototype_lup,
  parent_ns_ls
)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION}

\item{class_slots}{PARAM_DESCRIPTION}

\item{proto_ls}{Proto (a list)}

\item{prototype_lup}{Prototype (a lookup table)}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
create_ready_helper() is a Create function that creates a new R object. Specifically, this function implements an algorithm to create ready helper.NA
}
\keyword{internal}
