% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_create.R
\name{create_ready_show_mthd}
\alias{create_ready_show_mthd}
\title{Create ready show mthd}
\usage{
create_ready_show_mthd(class_name, meaningful_names)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{meaningful_names}{PARAM_DESCRIPTION}
}
\description{
create_ready_show_mthd() is a Create function that creates a new R object. Specifically, this function implements an algorithm to create ready show mthd. The function is called for its side effects and does not return a value.
}
\keyword{internal}
