% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_delete.R
\name{delete_files}
\alias{delete_files}
\title{Delete files}
\usage{
delete_files(dir_chr, pattern_chr)
}
\arguments{
\item{dir_chr}{Directory (a character vector of length 1)}

\item{pattern_chr}{Pattern (a character vector of length 1)}
}
\description{
delete_files() is a Delete function that deletes a file from a specified location. Specifically, this function implements an algorithm to delete files. The function is called for its side effects and does not return a value.
}
\keyword{internal}
