% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_delete.R
\name{delete_getters_setters}
\alias{delete_getters_setters}
\title{Delete getters setters}
\usage{
delete_getters_setters(x, output_dir)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{output_dir}{PARAM_DESCRIPTION}
}
\description{
delete_getters_setters() is a Delete function that deletes a file from a specified location. Specifically, this function implements an algorithm to delete getters setters. The function is called for its side effects and does not return a value.
}
\keyword{internal}
