% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exhibit.R
\name{exhibit.ready4class_pt_lup}
\alias{exhibit.ready4class_pt_lup}
\alias{exhibit,ready4class_pt_lup-method}
\title{Exhibit features of a dataset by printing them to the R console}
\usage{
exhibit.ready4class_pt_lup(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  use_lbls_as_col_nms_1L_lgl = T
)

\S4method{exhibit}{ready4class_pt_lup}(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  use_lbls_as_col_nms_1L_lgl = T
)
}
\arguments{
\item{x}{An instance of Class prototype lookup table}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: NULL}

\item{mkdn_tbl_ref_1L_chr}{Markdown table reference (a character vector of length one), Default: NULL}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{use_lbls_as_col_nms_1L_lgl}{Use labels as column names (a logical vector of length one), Default: T}
}
\description{
exhibit.ready4class_pt_lup() is an exhibit method that exhibits features of a class instance by printing to console. This method is implemented for the Class prototype lookup table The function is called for its side effects and does not return a value.
}
