% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_gen.R
\name{gen_pt_value_chr}
\alias{gen_pt_value_chr}
\title{Gen prototype value}
\usage{
gen_pt_value_chr(
  type_namespace = "",
  function_to_call = "",
  default_value = "",
  namespace_contexts = c("base")
)
}
\arguments{
\item{type_namespace}{PARAM_DESCRIPTION, Default: ''}

\item{function_to_call}{PARAM_DESCRIPTION, Default: ''}

\item{default_value}{PARAM_DESCRIPTION, Default: ''}

\item{namespace_contexts}{PARAM_DESCRIPTION, Default: c("base")}
}
\description{
gen_pt_value_chr() is a Gen function that generates values for an object. Specifically, this function implements an algorithm to gen a prototype value.NA
}
\keyword{internal}
