% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_generics_lup_tb.R
\docType{data}
\name{generics_lup_tb}
\alias{generics_lup_tb}
\title{Generics lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 5 columns.
}
\source{
\url{https://readyforwhatsnext.github.io/readyforwhatsnext/}
}
\usage{
generics_lup_tb
}
\description{
A lookup table to find descriptions of generics exported with the ready4class package suite.
}
\details{
A tibble

\describe{
  \item{fn_type_nm_chr}{Function type name (a character vector)}
  \item{fn_type_desc_chr}{Function type description (a character vector)}
  \item{first_arg_desc_chr}{First argument description (a character vector)}
  \item{second_arg_desc_chr}{Second argument description (a character vector)}
  \item{is_generic_lgl}{Is generic (a logical vector)}
}
}
\keyword{datasets}
