% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_class_files_chr}
\alias{get_class_files_chr}
\title{Get class files}
\usage{
get_class_files_chr(class_names_chr_vec, s3_lgl = T, output_dir_chr = NA)
}
\arguments{
\item{class_names_chr_vec}{Class names (a character vector)}

\item{s3_lgl}{S3 (a logical vector of length 1), Default: T}

\item{output_dir_chr}{Output directory (a character vector of length 1), Default: NA}
}
\description{
get_class_files_chr() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get class files. Function argument class_names_chr_vec specifies the where to look for the required object.NA
}
\keyword{internal}
