% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_class_fl_nms}
\alias{get_class_fl_nms}
\title{Get class file names}
\usage{
get_class_fl_nms(class_names_chr_vec, s3_lgl = T, output_dir_chr = NA)
}
\arguments{
\item{class_names_chr_vec}{PARAM_DESCRIPTION}

\item{s3_lgl}{S3 (a logical vector), Default: T}

\item{output_dir_chr}{Output directory (a character vector), Default: NA}
}
\description{
get_class_fl_nms() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get a class file names. Function argument class_names_chr_vec specifies the where to look for the required object.The function is called for its side effects and does not return a value.
}
\keyword{internal}
