% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_included_classes_chr_vec}
\alias{get_included_classes_chr_vec}
\title{Get included classes}
\usage{
get_included_classes_chr_vec(
  parent_chr,
  parent_ns_ls,
  prespecified_includes_chr = NULL
)
}
\arguments{
\item{parent_chr}{Parent (a character vector of length 1)}

\item{parent_ns_ls}{Parent namespace (a list)}

\item{prespecified_includes_chr}{Prespecified includes (a character vector of length 1), Default: NULL}
}
\description{
get_included_classes_chr_vec() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get included classes. Function argument parent_chr specifies the where to look for the required object.NA
}
\keyword{internal}
