% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_nms_of_curr_gnrcs}
\alias{get_nms_of_curr_gnrcs}
\title{Get names of current generics}
\usage{
get_nms_of_curr_gnrcs(req_pkgs_chr, generic_1L_chr)
}
\arguments{
\item{req_pkgs_chr}{Require packages (a character vector)}

\item{generic_1L_chr}{Generic (a character vector of length one)}
}
\value{
Names of current generics (a list)
}
\description{
get_nms_of_curr_gnrcs() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get names of current generics. The function returns Names of current generics (a list).
}
\keyword{internal}
