% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parent_cls_pt_fn}
\alias{get_parent_cls_pt_fn}
\title{Get parent class prototype}
\usage{
get_parent_cls_pt_fn(parent_chr, prototype_lup)
}
\arguments{
\item{parent_chr}{Parent (a character vector)}

\item{prototype_lup}{Prototype (a lookup table)}
}
\description{
get_parent_cls_pt_fn() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get parent cls a prototype. Function argument parent_chr specifies the where to look for the required object.The function is called for its side effects and does not return a value.
}
