% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parent_cls_pts}
\alias{get_parent_cls_pts}
\title{Get parent class prototypes}
\usage{
get_parent_cls_pts(parent_chr, parent_ns_ls, slot_names_chr_vec)
}
\arguments{
\item{parent_chr}{Parent (a character vector)}

\item{parent_ns_ls}{Parent namespace (a list)}

\item{slot_names_chr_vec}{PARAM_DESCRIPTION}
}
\description{
get_parent_cls_pts() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get parent cls prototypes. Function argument parent_chr specifies the where to look for the required object.The function is called for its side effects and does not return a value.
}
\keyword{internal}
