% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parent_ns_ls}
\alias{get_parent_ns_ls}
\title{Get parent namespace}
\usage{
get_parent_ns_ls(prototype_lup, parent_chr, dev_pckg_ns_chr)
}
\arguments{
\item{prototype_lup}{Prototype (a lookup table)}

\item{parent_chr}{Parent (a character vector of length 1)}

\item{dev_pckg_ns_chr}{Dev pckg namespace (a character vector of length 1)}
}
\description{
get_parent_ns_ls() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get parent a namespace. Function argument prototype_lup specifies the where to look for the required object.NA
}
\keyword{internal}
