% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parent_prototypes}
\alias{get_parent_prototypes}
\title{Get parent prototypes}
\usage{
get_parent_prototypes(parent_chr, parent_ns_ls, slot_names_chr_vec)
}
\arguments{
\item{parent_chr}{Parent (a character vector of length 1)}

\item{parent_ns_ls}{Parent namespace (a list)}

\item{slot_names_chr_vec}{Slot names (a character vector)}
}
\description{
get_parent_prototypes() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get parent prototypes. Function argument parent_chr specifies the where to look for the required object.NA
}
\keyword{internal}
