% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parent_slot_names}
\alias{get_parent_slot_names}
\title{Get parent slot names}
\usage{
get_parent_slot_names(parent_chr, parent_ns_ls)
}
\arguments{
\item{parent_chr}{Parent (a character vector of length 1)}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
get_parent_slot_names() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get parent slot names. Function argument parent_chr specifies the where to look for the required object. The function is called for its side effects and does not return a value.
}
\keyword{internal}
