% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_proto_list}
\alias{get_proto_list}
\title{Get proto list}
\usage{
get_proto_list(
  class_slots,
  type = NULL,
  values = NULL,
  make_val_string = TRUE,
  prototype_lup
)
}
\arguments{
\item{class_slots}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: NULL}

\item{values}{PARAM_DESCRIPTION, Default: NULL}

\item{make_val_string}{PARAM_DESCRIPTION, Default: TRUE}

\item{prototype_lup}{Prototype (a lookup table)}
}
\description{
get_proto_list() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get proto a list. Function argument class_slots specifies the where to look for the required object.NA
}
\keyword{internal}
