% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_make_tb.R
\name{is_ready4_class_make_tb}
\alias{is_ready4_class_make_tb}
\title{Class Make Table readyforwhatsnext S3 class}
\usage{
is_ready4_class_make_tb(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the
}
\description{
Check whether an object is a valid instance of the
}
\details{
MAKE CLASS TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
