% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{is_ready4_class_pt_lup}
\alias{is_ready4_class_pt_lup}
\title{NA}
\usage{
is_ready4_class_pt_lup(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4_class_pt_lup class
}
\description{
Check whether an object is a valid instance of the S3 class: ready4_class_pt_lup
}
\details{
PROTOTYPE LOOKUP Readyforwhatsnext S3 class of metadata of prototype classes.
}
