% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_constructor_tbl.R
\name{is_ready4_constructor_tbl}
\alias{is_ready4_constructor_tbl}
\title{Is CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class}
\usage{
is_ready4_constructor_tbl(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\description{
Check whether an object is a valid instance of the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\details{
CLASS CONSTRUCTOR TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
