% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor.R
\name{is_ready4class_constructor}
\alias{is_ready4class_constructor}
\title{is ready4class package constructor ready4 S3 class Constructor Table.}
\usage{
is_ready4class_constructor(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class Constructor Table.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class Constructor Table.
}
\details{
ready4 S3 class Constructor Table.
}
