% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor_tbl.R
\name{is_ready4class_constructor_tbl}
\alias{is_ready4class_constructor_tbl}
\title{Is ready4class package constructor table ready4 S3 class Constructor Table of metadata required to make new classes.}
\usage{
is_ready4class_constructor_tbl(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class Constructor Table of metadata required to make new classes.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class Constructor Table of metadata required to make new classes.
}
\details{
ready4 S3 class Constructor Table.
}
