% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_manifest.R
\name{is_ready4class_manifest}
\alias{is_ready4class_manifest}
\title{is ready4class package manifest ready4 S3 class Manifest.}
\usage{
is_ready4class_manifest(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class Manifest.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class Manifest.
}
\details{
ready4 S3 class Manifest.
}
