% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_pt_lup.R
\name{is_ready4class_pt_lup}
\alias{is_ready4class_pt_lup}
\title{Is ready4class package prototype lookup table ready4 S3 class Prototype Lookup Table of class metadata.}
\usage{
is_ready4class_pt_lup(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class Prototype Lookup Table of class metadata.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class Prototype Lookup Table of class metadata.
}
\details{
ready4 S3 class Prototype Lookup Table of class metadata.
}
