% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make.ready4class_constructor_tbl}
\alias{make.ready4class_constructor_tbl}
\title{Make method applied to ready4 S3 class Constructor Table of metadata required to make new classes..}
\usage{
make.ready4class_constructor_tbl(x, dev_pkg_ns_1L_chr, prefix_1L_chr)
}
\arguments{
\item{x}{An object}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}
}
\value{
Instance (ready4 S3 class Prototype Lookup Table of class metadata.)
}
\description{
make.ready4class_constructor_tbl() is a Make method that creates a new R object. This method is implemented for the ready4 S3 class Constructor Table of metadata required to make new classes. The function returns Instance (ready4 S3 class Prototype Lookup Table of class metadata.).
}
