% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_gen_ref_to_cls}
\alias{make_alg_to_gen_ref_to_cls}
\title{Make algorithm to generate reference to class}
\usage{
make_alg_to_gen_ref_to_cls(class_chr, package_chr = ".GlobalEnv")
}
\arguments{
\item{class_chr}{Class (a character vector)}

\item{package_chr}{Package (a character vector), Default: '.GlobalEnv'}
}
\description{
make_alg_to_gen_ref_to_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make alg to gen ref to cls.The function is called for its side effects and does not return a value.
}
