% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_get_pt_val}
\alias{make_alg_to_get_pt_val}
\title{Make algorithm to get prototype value}
\usage{
make_alg_to_get_pt_val(
  pt_ns_1L_chr = "",
  fn_to_call_1L_chr = "",
  default_val_1L_chr = "",
  attached_nss_chr = c("base")
)
}
\arguments{
\item{pt_ns_1L_chr}{Prototype namespace (a character vector of length one), Default: ''}

\item{fn_to_call_1L_chr}{Function to call (a character vector of length one), Default: ''}

\item{default_val_1L_chr}{Default value (a character vector of length one), Default: ''}

\item{attached_nss_chr}{Attached namespaces (a character vector), Default: c("base")}
}
\value{
Alg to get prototype val (a character vector of length one)
}
\description{
make_alg_to_get_pt_val() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make algorithm to get prototype value.The function returns alg to get prototype val (a character vector of length one).
}
