% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_set_gnrc}
\alias{make_alg_to_set_gnrc}
\title{Make algorithm to set generic}
\usage{
make_alg_to_set_gnrc(
  name_chr,
  args_chr_vec = c("x"),
  signature_chr = NA_character_,
  where_chr = NA_character_
)
}
\arguments{
\item{name_chr}{Name (a character vector)}

\item{args_chr_vec}{PARAM_DESCRIPTION, Default: c("x")}

\item{signature_chr}{Signature (a character vector), Default: 'NA'}

\item{where_chr}{Where (a character vector), Default: 'NA'}
}
\description{
make_alg_to_set_gnrc() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make alg to set gnrc.The function is called for its side effects and does not return a value.
}
