% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_set_mthd}
\alias{make_alg_to_set_mthd}
\title{Make algorithm to set method}
\usage{
make_alg_to_set_mthd(
  name_chr,
  class_chr,
  fn,
  package_chr = NA_character_,
  where_chr = NA_character_
)
}
\arguments{
\item{name_chr}{Name (a character vector)}

\item{class_chr}{Class (a character vector)}

\item{fn}{Function (a function)}

\item{package_chr}{Package (a character vector), Default: 'NA'}

\item{where_chr}{Where (a character vector), Default: 'NA'}
}
\description{
make_alg_to_set_mthd() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make alg to set mthd.The function is called for its side effects and does not return a value.
}
