% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_set_validity_of_r4_cls}
\alias{make_alg_to_set_validity_of_r4_cls}
\title{Make algorithm to set validity of readyforwhatsnext S4 class}
\usage{
make_alg_to_set_validity_of_r4_cls(
  class_name,
  parent,
  not_same_length = NULL,
  allowed_values = NULL,
  names_include = NULL,
  print_validator = FALSE
)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION}

\item{not_same_length}{PARAM_DESCRIPTION, Default: NULL}

\item{allowed_values}{PARAM_DESCRIPTION, Default: NULL}

\item{names_include}{PARAM_DESCRIPTION, Default: NULL}

\item{print_validator}{PARAM_DESCRIPTION, Default: FALSE}
}
\description{
make_alg_to_set_validity_of_r4_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make alg to set validity of a readyforwhatsnext S4 cls.The function is called for its side effects and does not return a value.
}
