% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_write_gtr_str_mthds}
\alias{make_alg_to_write_gtr_str_mthds}
\title{Make algorithm to write getter setter methods}
\usage{
make_alg_to_write_gtr_str_mthds(
  class_name,
  parent,
  print_accessors,
  output_folder,
  ignore_ns_chr,
  required_pckg_chr_vec,
  parent_ns_ls
)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION}

\item{print_accessors}{PARAM_DESCRIPTION}

\item{output_folder}{PARAM_DESCRIPTION}

\item{ignore_ns_chr}{Ignore namespace (a character vector)}

\item{required_pckg_chr_vec}{PARAM_DESCRIPTION}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
make_alg_to_write_gtr_str_mthds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make alg to write a getter setter mthds.The function is called for its side effects and does not return a value.
}
\keyword{internal}
