% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_and_tf_curr_gen_ls}
\alias{make_and_tf_curr_gen_ls}
\title{Make and tf curr gen}
\usage{
make_and_tf_curr_gen_ls(required_pckg_chr_vec, generic_chr, ignore_ns_chr)
}
\arguments{
\item{required_pckg_chr_vec}{Required pckg (a character vector)}

\item{generic_chr}{Generic (a character vector of length 1)}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1)}
}
\description{
make_and_tf_curr_gen_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make and tf curr gen. The function is called for its side effects and does not return a value.
}
\keyword{internal}
