% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_and_update.R
\name{make_and_update.ready4_class_make_tb}
\alias{make_and_update.ready4_class_make_tb}
\title{Make and update method applied to Class Make Table readyforwhatsnext S3 class.}
\usage{
\method{make_and_update}{ready4_class_make_tb}(
  x,
  dev_pkg_ns_1L_chr = ready4fun::get_dev_pkg_nm(),
  name_pfx_1L_chr = paste0(ready4fun::get_dev_pkg_nm(), "_"),
  output_dir_1L_chr = "R",
  delete_cdn_ptrn_chr = NA_character_,
  file_exists_cdn_1L_chr = "overwrite",
  init_class_pt_lup = NULL,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop"
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one), Default: ready4fun::get_dev_pkg_nm()}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one), Default: paste0(ready4fun::get_dev_pkg_nm(), "_")}

\item{output_dir_1L_chr}{Output directory (a character vector of length one), Default: 'R'}

\item{delete_cdn_ptrn_chr}{Delete condition pattern (a character vector), Default: 'NA'}

\item{file_exists_cdn_1L_chr}{File exists condition (a character vector of length one), Default: 'overwrite'}

\item{init_class_pt_lup}{Init class prototype (a lookup table), Default: NULL}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector), Default: 'NA'}

\item{req_pkgs_chr}{Req packages (a character vector), Default: 'NA'}

\item{class_in_cache_cdn_1L_chr}{Class in cache condition (a character vector of length one), Default: 'stop'}
}
\value{
Inst of ready4 class prototype (a lookup table)
}
\description{
make_and_update.ready4_class_make_tb() is a Make and Update method that applies a Make method and then updates the output of that method. This method is implemented for the Class Make Table readyforwhatsnext S3 class.The function returns inst of ready4 class prototype (a lookup table).
}
\keyword{internal}
