% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_child_fn_chr}
\alias{make_child_fn_chr}
\title{Make child function}
\usage{
make_child_fn_chr(child_ext_fn_chr, parent_chr, prototype_lup, prepend_lgl = T)
}
\arguments{
\item{child_ext_fn_chr}{Child ext function (a character vector of length 1)}

\item{parent_chr}{Parent (a character vector of length 1)}

\item{prototype_lup}{Prototype (a lookup table)}

\item{prepend_lgl}{Prepend (a logical vector of length 1), Default: T}
}
\description{
make_child_fn_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make child a function. The function is called for its side effects and does not return a value.
}
\keyword{internal}
