% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_class_mk_r3}
\alias{make_class_mk_r3}
\title{Make class mk}
\usage{
make_class_mk_r3(class_mk_ls)
}
\arguments{
\item{class_mk_ls}{Class mk (a list)}
}
\description{
make_class_mk_r3() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make class mk.NA
}
\keyword{internal}
