% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_class_mk_tb}
\alias{make_class_mk_tb}
\title{Make class mk}
\usage{
make_class_mk_tb(class_mk_ls)
}
\arguments{
\item{class_mk_ls}{Class mk (a list)}
}
\description{
make_class_mk_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make class mk. The function is called for its side effects and does not return a value.
}
\keyword{internal}
