% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_class_pts_tb}
\alias{make_class_pts_tb}
\title{Make class prototypes tibble}
\usage{
make_class_pts_tb(class_mk_ls)
}
\arguments{
\item{class_mk_ls}{Class make (a list)}
}
\value{
Class prototypes (a tibble)
}
\description{
make_class_pts_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make class prototypes tibble. The function returns Class prototypes (a tibble).
}
