% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_class_type_mk_tb}
\alias{make_class_type_mk_tb}
\title{Make class type mk}
\usage{
make_class_type_mk_tb(class_type_mk_ls, make_s3_lgl = T)
}
\arguments{
\item{class_type_mk_ls}{Class type mk (a list)}

\item{make_s3_lgl}{Make S3 (a logical vector of length 1), Default: T}
}
\description{
make_class_type_mk_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make class type mk. The function is called for its side effects and does not return a value.
}
\keyword{internal}
