% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{make_classes}
\alias{make_classes}
\title{Make classes}
\usage{
make_classes(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\description{
make_classes() is a Make Classes generic that writes new classes.
}
\keyword{internal}
