% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_classes.R
\name{make_classes.ready4_class_make_tb}
\alias{make_classes.ready4_class_make_tb}
\title{Make classes method applied to Class Make Table readyforwhatsnext S3 class.}
\usage{
\method{make_classes}{ready4_class_make_tb}(
  x,
  name_pfx_1L_chr,
  output_dir_1L_chr,
  file_exists_cdn_1L_chr = NULL,
  prototype_lup = NULL,
  nss_to_ignore_chr,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop"
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one)}

\item{file_exists_cdn_1L_chr}{File exists condition (a character vector of length one), Default: NULL}

\item{prototype_lup}{Prototype (a lookup table), Default: NULL}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector)}

\item{req_pkgs_chr}{Req packages (a character vector), Default: 'NA'}

\item{class_in_cache_cdn_1L_chr}{Class in cache condition (a character vector of length one), Default: 'stop'}
}
\description{
make_classes.ready4_class_make_tb() is a Make Classes method that writes new classes. This method is implemented for the Class Make Table readyforwhatsnext S3 class.NA
}
