% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_className_chr}
\alias{make_className_chr}
\title{Make className}
\usage{
make_className_chr(class_chr, package_chr = ".GlobalEnv")
}
\arguments{
\item{class_chr}{Class (a character vector of length 1)}

\item{package_chr}{Package (a character vector of length 1), Default: '.GlobalEnv'}
}
\description{
make_className_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make className. The function is called for its side effects and does not return a value.
}
\keyword{internal}
