% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_current_generics_ls}
\alias{make_current_generics_ls}
\title{Make current generics}
\usage{
make_current_generics_ls(required_pckg_chr_vec, generic_chr)
}
\arguments{
\item{required_pckg_chr_vec}{Required pckg (a character vector)}

\item{generic_chr}{Generic (a character vector of length 1)}
}
\description{
make_current_generics_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make current generics. The function is called for its side effects and does not return a value.
}
\keyword{internal}
