% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_dmt_inc_tag}
\alias{make_dmt_inc_tag}
\title{Make documentation inc tag}
\usage{
make_dmt_inc_tag(class_names_chr_vec, s3_lgl = T)
}
\arguments{
\item{class_names_chr_vec}{PARAM_DESCRIPTION}

\item{s3_lgl}{S3 (a logical vector), Default: T}
}
\description{
make_dmt_inc_tag() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make documentation inc tag.The function is called for its side effects and does not return a value.
}
