% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_pt_to_check_r3_cls_inhtc}
\alias{make_fn_pt_to_check_r3_cls_inhtc}
\title{Make function prototype to check readyforwhatsnext S3 class inheritance}
\usage{
make_fn_pt_to_check_r3_cls_inhtc(class_nm_1L_chr, s3_validator_ls)
}
\arguments{
\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{s3_validator_ls}{S3 validator (a list)}
}
\value{
Function prototype to check readyforwhatsnext S3 class (an inheritance)
}
\description{
make_fn_pt_to_check_r3_cls_inhtc() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function prototype to check readyforwhatsnext s3 class inheritance. The function is called for its side effects and does not return a value.
}
\keyword{internal}
