% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_pt_to_make_r3_cls_pt}
\alias{make_fn_pt_to_make_r3_cls_pt}
\title{Make function prototype to make readyforwhatsnext S3 class prototype}
\usage{
make_fn_pt_to_make_r3_cls_pt(
  type_1L_chr,
  pt_ns_1L_chr,
  vals_ls,
  ordered_1L_lgl,
  class_nm_1L_chr,
  parent_cls_nm_1L_chr,
  prototype_lup
)
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one)}

\item{pt_ns_1L_chr}{Prototype namespace (a character vector of length one)}

\item{vals_ls}{Values (a list)}

\item{ordered_1L_lgl}{Ordered (a logical vector of length one)}

\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{prototype_lup}{Prototype (a lookup table)}
}
\description{
make_fn_pt_to_make_r3_cls_pt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make function prototype to a make readyforwhatsnext S3 cls a prototype.The function is called for its side effects and does not return a value.
}
\keyword{internal}
