% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_pt_to_make_unvld_r3_cls_inst}
\alias{make_fn_pt_to_make_unvld_r3_cls_inst}
\title{Make function prototype to make unvld readyforwhatsnext S3 class instance}
\usage{
make_fn_pt_to_make_unvld_r3_cls_inst(
  type_1L_chr,
  pt_chkr_pfx_1L_chr,
  pt_ns_1L_chr,
  class_nm_1L_chr,
  s3_prototype_ls
)
}
\arguments{
\item{type_1L_chr}{Type (a character vector of length one)}

\item{pt_chkr_pfx_1L_chr}{Prototype checker prefix (a character vector of length one)}

\item{pt_ns_1L_chr}{Prototype namespace (a character vector of length one)}

\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{s3_prototype_ls}{S3 prototype (a list)}
}
\value{
Function prototype to make unvld readyforwhatsnext S3 class (an instance)
}
\description{
make_fn_pt_to_make_unvld_r3_cls_inst() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function prototype to make unvld readyforwhatsnext s3 class instance. The function is called for its side effects and does not return a value.
}
