% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fn_pt_to_make_vldd_r3_cls_inst}
\alias{make_fn_pt_to_make_vldd_r3_cls_inst}
\title{Make function prototype to make validated readyforwhatsnext S3 class instance}
\usage{
make_fn_pt_to_make_vldd_r3_cls_inst(
  class_nm_1L_chr,
  s3_prototype_ls,
  s3_constructor_ls,
  s3_validator_ls
)
}
\arguments{
\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{s3_prototype_ls}{S3 prototype (a list)}

\item{s3_constructor_ls}{S3 constructor (a list)}

\item{s3_validator_ls}{S3 validator (a list)}
}
\description{
make_fn_pt_to_make_vldd_r3_cls_inst() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make function prototype to a make vldd a readyforwhatsnext S3 cls inst.The function is called for its side effects and does not return a value.
}
