% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gen_fn_chr}
\alias{make_gen_fn_chr}
\title{Make gen function}
\usage{
make_gen_fn_chr(name_chr, args_chr_vec)
}
\arguments{
\item{name_chr}{Name (a character vector of length 1)}

\item{args_chr_vec}{Arguments (a character vector)}
}
\description{
make_gen_fn_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make gen a function. The function is called for its side effects and does not return a value.
}
\keyword{internal}
