% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_generic_chr}
\alias{make_generic_chr}
\title{Make generic}
\usage{
make_generic_chr(
  name_chr,
  args_chr_vec = c("x"),
  signature_chr = NA_character_,
  where_chr = NA_character_
)
}
\arguments{
\item{name_chr}{Name (a character vector of length 1)}

\item{args_chr_vec}{Arguments (a character vector), Default: c("x")}

\item{signature_chr}{Signature (a character vector of length 1), Default: 'NA'}

\item{where_chr}{Where (a character vector of length 1), Default: 'NA'}
}
\description{
make_generic_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make generic.NA
}
\keyword{internal}
