% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gnrc_fn}
\alias{make_gnrc_fn}
\title{Make generic}
\usage{
make_gnrc_fn(name_chr, args_chr_vec)
}
\arguments{
\item{name_chr}{Name (a character vector)}

\item{args_chr_vec}{PARAM_DESCRIPTION}
}
\description{
make_gnrc_fn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make gnrc.The function is called for its side effects and does not return a value.
}
\keyword{internal}
