% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gnrc_mthd_pair_ls}
\alias{make_gnrc_mthd_pair_ls}
\title{Make generic method pair}
\usage{
make_gnrc_mthd_pair_ls(
  name_chr,
  args_chr_vec = c("x"),
  signature_chr = NA_character_,
  package_chr = NA_character_,
  where_chr = NA_character_,
  class_chr,
  fn
)
}
\arguments{
\item{name_chr}{Name (a character vector)}

\item{args_chr_vec}{PARAM_DESCRIPTION, Default: c("x")}

\item{signature_chr}{Signature (a character vector), Default: 'NA'}

\item{package_chr}{Package (a character vector), Default: 'NA'}

\item{where_chr}{Where (a character vector), Default: 'NA'}

\item{class_chr}{Class (a character vector)}

\item{fn}{Function (a function)}
}
\description{
make_gnrc_mthd_pair_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make gnrc mthd pair.The function is called for its side effects and does not return a value.
}
