% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_helper_fn}
\alias{make_helper_fn}
\title{Make helper}
\usage{
make_helper_fn(
  class_name,
  parent,
  class_slots,
  proto_ls,
  prototype_lup,
  parent_ns_ls
)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION}

\item{class_slots}{PARAM_DESCRIPTION}

\item{proto_ls}{Proto (a list)}

\item{prototype_lup}{Prototype (a lookup table)}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
make_helper_fn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make helper.The function is called for its side effects and does not return a value.
}
\keyword{internal}
