% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_import_packages_ls}
\alias{make_import_packages_ls}
\title{Make import packages}
\usage{
make_import_packages_ls(current_generics_ls, fn_name_chr, ignore_ns_chr)
}
\arguments{
\item{current_generics_ls}{Current generics (a list)}

\item{fn_name_chr}{Function name (a character vector of length 1)}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1)}
}
\description{
make_import_packages_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make an import packages. The function is called for its side effects and does not return a value.
}
\keyword{internal}
