% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_incld_fn_fls}
\alias{make_incld_fn_fls}
\title{Make included function files}
\usage{
make_incld_fn_fls(vals_ls, fns_env_ls)
}
\arguments{
\item{vals_ls}{Values (a list)}

\item{fns_env_ls}{Functions (a list of environments)}
}
\value{
Function files (a character vector)
}
\description{
make_incld_fn_fls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make included function files. The function returns Function files (a character vector).
}
\keyword{internal}
