% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_include_tag_chr}
\alias{make_include_tag_chr}
\title{Make include tag}
\usage{
make_include_tag_chr(class_names_chr_vec, s3_lgl = T)
}
\arguments{
\item{class_names_chr_vec}{Class names (a character vector)}

\item{s3_lgl}{S3 (a logical vector of length 1), Default: T}
}
\description{
make_include_tag_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make include tag.NA
}
\keyword{internal}
