% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_lines_for_writing_dmtd_fn}
\alias{make_lines_for_writing_dmtd_fn}
\title{Make lines for writing documented}
\usage{
make_lines_for_writing_dmtd_fn(
  fn_name,
  fn_text,
  fn_type,
  class_name,
  class_desc
)
}
\arguments{
\item{fn_name}{PARAM_DESCRIPTION}

\item{fn_text}{PARAM_DESCRIPTION}

\item{fn_type}{PARAM_DESCRIPTION}

\item{class_name}{PARAM_DESCRIPTION}

\item{class_desc}{PARAM_DESCRIPTION}
}
\description{
make_lines_for_writing_dmtd_fn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make lines for writing documented.The function is called for its side effects and does not return a value.
}
