% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ls_of_pkgs_to_imp}
\alias{make_ls_of_pkgs_to_imp}
\title{Make list of packages to import}
\usage{
make_ls_of_pkgs_to_imp(current_generics_ls, fn_name_chr, ignore_ns_chr)
}
\arguments{
\item{current_generics_ls}{Current generics (a list)}

\item{fn_name_chr}{Function name (a character vector)}

\item{ignore_ns_chr}{Ignore namespace (a character vector)}
}
\description{
make_ls_of_pkgs_to_imp() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make list of packages to an import.The function is called for its side effects and does not return a value.
}
