% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ls_of_tfd_nms_of_curr_gnrcs}
\alias{make_ls_of_tfd_nms_of_curr_gnrcs}
\title{Make list of transformed names of current generics}
\usage{
make_ls_of_tfd_nms_of_curr_gnrcs(
  required_pckg_chr_vec,
  generic_chr,
  ignore_ns_chr
)
}
\arguments{
\item{required_pckg_chr_vec}{PARAM_DESCRIPTION}

\item{generic_chr}{Generic (a character vector)}

\item{ignore_ns_chr}{Ignore namespace (a character vector)}
}
\description{
make_ls_of_tfd_nms_of_curr_gnrcs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make list of tfd names of curr gnrcs.The function is called for its side effects and does not return a value.
}
\keyword{internal}
