% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_make_lup.R
\name{make_lup}
\alias{make_lup}
\alias{make_lup.ready4class_constructor_tbl}
\alias{make_lup,ready4class_constructor_tbl-method}
\title{Make lookup table}
\usage{
make_lup(x, ...)

make_lup.ready4class_constructor_tbl(x, dev_pkg_ns_1L_chr, prefix_1L_chr)

\S4method{make_lup}{ready4class_constructor_tbl}(x, dev_pkg_ns_1L_chr, prefix_1L_chr)
}
\arguments{
\item{x}{An instance of ready4 S3 class Constructor Table of metadata required to make new classes.}

\item{...}{Additional arguments (an additional arguments)}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}
}
\value{
Instance (ready4 S3 class Prototype Lookup Table of class metadata.)
}
\description{
make_lup() is a Make Lookup Table generic that applies a Make method and then updates the output of that method.

make_lup.ready4class_constructor_tbl() is a Make Lookup Table method that applies a Make method and then updates the output of that method. This method is implemented for the ready4 S3 class Constructor Table of metadata required to make new classes.. The function returns Instance (ready4 S3 class Prototype Lookup Table of class metadata.).
}
