% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_lup.R
\name{make_lup.ready4_class_make_tb}
\alias{make_lup.ready4_class_make_tb}
\title{Make lookup table method applied to Class Make Table readyforwhatsnext S3 class.}
\usage{
\method{make_lup}{ready4_class_make_tb}(x, dev_pckg_namespace, prefix)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{dev_pckg_namespace}{PARAM_DESCRIPTION}

\item{prefix}{PARAM_DESCRIPTION}
}
\description{
make_lup.ready4_class_make_tb() is a Make Lookup Table method that makes a lookup table. This method is implemented for the Class Make Table readyforwhatsnext S3 class.NA
}
\keyword{internal}
