% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_lup.R
\name{make_lup.ready4_constructor_tbl}
\alias{make_lup.ready4_constructor_tbl}
\title{Make lookup table method applied to readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE.}
\usage{
\method{make_lup}{ready4_constructor_tbl}(x, dev_pkg_ns_1L_chr, prefix_1L_chr)
}
\arguments{
\item{x}{An instance of readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}
}
\value{
Instance (a readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE)
}
\description{
make_lup.ready4_constructor_tbl() is a Make Lookup Table method that applies a Make method and then updates the output of that method. This method is implemented for the readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE. The function returns Instance (a readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE).
}
\keyword{internal}
