% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_meth_fn_chr}
\alias{make_meth_fn_chr}
\title{Make meth function}
\usage{
make_meth_fn_chr(fn)
}
\arguments{
\item{fn}{Function (a function)}
}
\description{
make_meth_fn_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make meth a function.NA
}
\keyword{internal}
