% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_method_chr}
\alias{make_method_chr}
\title{Make method}
\usage{
make_method_chr(
  name_chr,
  class_chr,
  fn,
  package_chr = NA_character_,
  where_chr = NA_character_
)
}
\arguments{
\item{name_chr}{Name (a character vector of length 1)}

\item{class_chr}{Class (a character vector of length 1)}

\item{fn}{Function (a function)}

\item{package_chr}{Package (a character vector of length 1), Default: 'NA'}

\item{where_chr}{Where (a character vector of length 1), Default: 'NA'}
}
\description{
make_method_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make method. The function is called for its side effects and does not return a value.
}
\keyword{internal}
