% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_new_classes}
\alias{make_new_classes}
\title{Make new classes}
\usage{
make_new_classes(
  new_classes_ls,
  pckg_name_chr,
  class_pfx_chr,
  R_dir_chr = "R",
  pt_lup,
  description_ls = NULL,
  ignore_ns_chr = NA_character_,
  required_pckg_chr_vec = NA_character_
)
}
\arguments{
\item{new_classes_ls}{New classes (a list)}

\item{pckg_name_chr}{Pckg name (a character vector of length 1)}

\item{class_pfx_chr}{Class prefix (a character vector of length 1)}

\item{R_dir_chr}{R directory (a character vector of length 1), Default: 'R'}

\item{pt_lup}{Prototype (a lookup table)}

\item{description_ls}{Description (a list), Default: NULL}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1), Default: 'NA'}

\item{required_pckg_chr_vec}{Required pckg (a character vector), Default: 'NA'}
}
\description{
make_new_classes() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make new classes.NA
}
\keyword{internal}
