% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_make_tb.R
\name{make_new_ready4_class_make_tb}
\alias{make_new_ready4_class_make_tb}
\title{Class Make Table readyforwhatsnext S3 class}
\usage{
make_new_ready4_class_make_tb(x)
}
\arguments{
\item{x}{A prototype for the make_ready4_class_make_tb class}
}
\value{
An unvalidated instance of the
}
\description{
Create a new unvalidated instance of the
}
\details{
MAKE CLASS TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
