% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor.R
\name{make_new_ready4class_constructor}
\alias{make_new_ready4class_constructor}
\title{make new ready4class package constructor ready4 S3 class Constructor Table.}
\usage{
make_new_ready4class_constructor(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class Constructor Table.}
}
\value{
An unvalidated instance of the ready4 S3 class Constructor Table.
}
\description{
Create a new unvalidated instance of the ready4 S3 class Constructor Table.
}
\details{
ready4 S3 class Constructor Table.
}
\keyword{internal}
