% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor_tbl.R
\name{make_new_ready4class_constructor_tbl}
\alias{make_new_ready4class_constructor_tbl}
\title{Make new ready4class package constructor table ready4 S3 class Constructor Table of metadata required to make new classes.}
\usage{
make_new_ready4class_constructor_tbl(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class Constructor Table of metadata required to make new classes.}
}
\value{
An unvalidated instance of the ready4 S3 class Constructor Table of metadata required to make new classes.
}
\description{
Create a new unvalidated instance of the ready4 S3 class Constructor Table of metadata required to make new classes.
}
\details{
ready4 S3 class Constructor Table.
}
