% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_manifest.R
\name{make_new_ready4class_manifest}
\alias{make_new_ready4class_manifest}
\title{make new ready4class manifest Class creation manifest}
\usage{
make_new_ready4class_manifest(x)
}
\arguments{
\item{x}{A prototype for the Class creation manifest}
}
\value{
An unvalidated instance of the Class creation manifest
}
\description{
Create a new unvalidated instance of the Class creation manifest
}
\details{
Class creation manifest
}
\keyword{internal}
