% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_one_row_pt_tb_for_new_r3_cls}
\alias{make_one_row_pt_tb_for_new_r3_cls}
\title{Make one row prototype tibble for new ready4 S3 class}
\usage{
make_one_row_pt_tb_for_new_r3_cls(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
One row class prototype (a tibble)
}
\description{
make_one_row_pt_tb_for_new_r3_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make one row prototype tibble for new ready4 s3 class. The function returns One row class prototype (a tibble).
}
