% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_one_row_pt_tb_for_new_r4_cls}
\alias{make_one_row_pt_tb_for_new_r4_cls}
\title{Make one row prototype tibble for new readyforwhatsnext S4 class}
\usage{
make_one_row_pt_tb_for_new_r4_cls(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
One row class prototype (a tibble)
}
\description{
make_one_row_pt_tb_for_new_r4_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make one row a prototype tibble for new a readyforwhatsnext S4 cls.The function returns one row class prototype (a tibble).
}
\keyword{internal}
