% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_make_tb.R
\name{make_prototype_ready4_class_make_tb}
\alias{make_prototype_ready4_class_make_tb}
\title{NA}
\usage{
make_prototype_ready4_class_make_tb()
}
\value{
A prototype for ready4_class_make_tb class
}
\description{
Create a new prototype for S3 class: ready4_class_make_tb
}
\details{
MAKE CLASS TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
