% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{make_prototype_ready4_class_pt_lup}
\alias{make_prototype_ready4_class_pt_lup}
\title{Make prototype readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE}
\usage{
make_prototype_ready4_class_pt_lup()
}
\value{
A prototype for readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE
}
\description{
Create a new prototype for the readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE
}
\details{
Readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE class of metadata of prototype classes.
}
