% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_constructor_tbl.R
\name{make_prototype_ready4_constructor_tbl}
\alias{make_prototype_ready4_constructor_tbl}
\title{Make prototype readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE}
\usage{
make_prototype_ready4_constructor_tbl()
}
\value{
A prototype for readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE
}
\description{
Create a new prototype for the readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE
}
\details{
Readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE of metadata required to make new classes.
}
