% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pt_ls}
\alias{make_pt_ls}
\title{Make prototype}
\usage{
make_pt_ls(
  class_slots,
  type = NULL,
  values = NULL,
  make_val_string = TRUE,
  prototype_lup
)
}
\arguments{
\item{class_slots}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: NULL}

\item{values}{PARAM_DESCRIPTION, Default: NULL}

\item{make_val_string}{PARAM_DESCRIPTION, Default: TRUE}

\item{prototype_lup}{Prototype (a lookup table)}
}
\description{
make_pt_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make prototype.The function is called for its side effects and does not return a value.
}
