% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pt_ls_for_new_r3_cls}
\alias{make_pt_ls_for_new_r3_cls}
\title{Make prototype for new readyforwhatsnext S3 class}
\usage{
make_pt_ls_for_new_r3_cls(
  class_name_chr,
  type_chr,
  type_ns_chr,
  type_checker_pfx_chr,
  values_chr,
  ordered_lgl,
  parent_chr,
  prototype_lup,
  min_max_values_dbl_vec,
  start_end_values_dbl_vec,
  ignore_ns_chr
)
}
\arguments{
\item{class_name_chr}{Class name (a character vector)}

\item{type_chr}{Type (a character vector)}

\item{type_ns_chr}{Type namespace (a character vector)}

\item{type_checker_pfx_chr}{Type checker prefix (a character vector)}

\item{values_chr}{Values (a character vector)}

\item{ordered_lgl}{Ordered (a logical vector)}

\item{parent_chr}{Parent (a character vector)}

\item{prototype_lup}{Prototype (a lookup table)}

\item{min_max_values_dbl_vec}{PARAM_DESCRIPTION}

\item{start_end_values_dbl_vec}{PARAM_DESCRIPTION}

\item{ignore_ns_chr}{Ignore namespace (a character vector)}
}
\description{
make_pt_ls_for_new_r3_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make prototype for new a readyforwhatsnext S3 cls.The function is called for its side effects and does not return a value.
}
\keyword{internal}
