% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{make_pt_ready4_class_pt_lup}
\alias{make_pt_ready4_class_pt_lup}
\title{Make prototype ready4 S3 class CLASS PROTOTYPE LOOKUP TABLE}
\usage{
make_pt_ready4_class_pt_lup(
  type_chr = character(0),
  val_chr = character(0),
  pt_ns_chr = character(0),
  fn_to_call_chr = character(0),
  default_val_chr = character(0),
  old_class_lgl = logical(0)
)
}
\arguments{
\item{type_chr}{Type (a character vector), Default: character(0)}

\item{val_chr}{Value (a character vector), Default: character(0)}

\item{pt_ns_chr}{Prototype namespace (a character vector), Default: character(0)}

\item{fn_to_call_chr}{Function to call (a character vector), Default: character(0)}

\item{default_val_chr}{Default value (a character vector), Default: character(0)}

\item{old_class_lgl}{Old class (a logical vector), Default: logical(0)}
}
\value{
A prototype for ready4 S3 class CLASS PROTOTYPE LOOKUP TABLE
}
\description{
Create a new prototype for the ready4 S3 class CLASS PROTOTYPE LOOKUP TABLE
}
\details{
ready4 S3 class CLASS PROTOTYPE LOOKUP TABLE class of metadata of prototype classes.
}
