% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_manifest.R
\name{make_pt_ready4class_manifest}
\alias{make_pt_ready4class_manifest}
\title{Make prototype ready4class package manifest ready4 S3 class Manifest.}
\usage{
make_pt_ready4class_manifest(
  pkg_setup_r3 = ready4fun::ready4fun_pkg_setup(),
  constructor_tbl_r3 = ready4class_constructor_tbl()
)
}
\arguments{
\item{pkg_setup_r3}{Package setup (a ready4 S3), Default: ready4fun::ready4fun_pkg_setup()}

\item{constructor_tbl_r3}{Constructor table (a ready4 S3), Default: ready4class_constructor_tbl()}
}
\value{
A prototype for ready4 S3 class Manifest.
}
\description{
Create a new prototype for the ready4 S3 class Manifest.
}
\details{
ready4 S3 class Manifest.
}
