% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pt_tb_for_new_r3_cls}
\alias{make_pt_tb_for_new_r3_cls}
\title{Make prototype tibble for new ready4 S3 class}
\usage{
make_pt_tb_for_new_r3_cls(x)
}
\arguments{
\item{x}{An object}
}
\value{
Prototype tibble for new ready4 S3 class (a tibble)
}
\description{
make_pt_tb_for_new_r3_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make prototype tibble for new ready4 s3 class. The function returns Prototype tibble for new ready4 S3 class (a tibble).
}
\keyword{internal}
