% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ready_s3}
\alias{make_ready_s3}
\title{Make ready}
\usage{
make_ready_s3(
  name_stub,
  name_prefix = "ready4_",
  output_folder = "data-raw",
  class_desc = "",
  parent = NULL,
  type,
  type_namespace = "",
  type_checker_prefix = "is.",
  values = NULL,
  ordered = FALSE,
  allowed_values = NULL,
  min_max_values = NULL,
  start_end_values = NULL,
  prototype_lup,
  ignore_ns_chr = NA_character_,
  file_exists_logic = "skip"
)
}
\arguments{
\item{name_stub}{PARAM_DESCRIPTION}

\item{name_prefix}{PARAM_DESCRIPTION, Default: 'ready4_'}

\item{output_folder}{PARAM_DESCRIPTION, Default: 'data-raw'}

\item{class_desc}{PARAM_DESCRIPTION, Default: ''}

\item{parent}{PARAM_DESCRIPTION, Default: NULL}

\item{type}{PARAM_DESCRIPTION}

\item{type_namespace}{PARAM_DESCRIPTION, Default: ''}

\item{type_checker_prefix}{PARAM_DESCRIPTION, Default: 'is.'}

\item{values}{PARAM_DESCRIPTION, Default: NULL}

\item{ordered}{PARAM_DESCRIPTION, Default: FALSE}

\item{allowed_values}{PARAM_DESCRIPTION, Default: NULL}

\item{min_max_values}{PARAM_DESCRIPTION, Default: NULL}

\item{start_end_values}{PARAM_DESCRIPTION, Default: NULL}

\item{prototype_lup}{Prototype (a lookup table)}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1), Default: 'NA'}

\item{file_exists_logic}{PARAM_DESCRIPTION, Default: 'skip'}
}
\description{
make_ready_s3() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ready.NA
}
\keyword{internal}
