% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ready_s3_checker}
\alias{make_ready_s3_checker}
\title{Make ready S3 checker}
\usage{
make_ready_s3_checker(class_name, s3_validator)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{s3_validator}{PARAM_DESCRIPTION}
}
\description{
make_ready_s3_checker() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ready a S3 checker.NA
}
\keyword{internal}
