% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ready_s3_constructor}
\alias{make_ready_s3_constructor}
\title{Make ready S3 constructor}
\usage{
make_ready_s3_constructor(
  type,
  type_checker_prefix,
  type_namespace,
  class_name,
  s3_prototype
)
}
\arguments{
\item{type}{PARAM_DESCRIPTION}

\item{type_checker_prefix}{PARAM_DESCRIPTION}

\item{type_namespace}{PARAM_DESCRIPTION}

\item{class_name}{PARAM_DESCRIPTION}

\item{s3_prototype}{PARAM_DESCRIPTION}
}
\description{
make_ready_s3_constructor() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ready a S3 constructor. The function is called for its side effects and does not return a value.
}
\keyword{internal}
