% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ready_s3_prototype}
\alias{make_ready_s3_prototype}
\title{Make ready S3 prototype}
\usage{
make_ready_s3_prototype(
  type,
  type_namespace,
  values,
  ordered,
  class_name,
  parent_chr,
  prototype_lup
)
}
\arguments{
\item{type}{PARAM_DESCRIPTION}

\item{type_namespace}{PARAM_DESCRIPTION}

\item{values}{PARAM_DESCRIPTION}

\item{ordered}{PARAM_DESCRIPTION}

\item{class_name}{PARAM_DESCRIPTION}

\item{parent_chr}{Parent (a character vector of length 1)}

\item{prototype_lup}{Prototype (a lookup table)}
}
\description{
make_ready_s3_prototype() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ready a S3 prototype.NA
}
\keyword{internal}
