% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ready_s3_valid_instance}
\alias{make_ready_s3_valid_instance}
\title{Make ready S3 valid instance}
\usage{
make_ready_s3_valid_instance(
  class_name,
  s3_prototype,
  s3_constructor,
  s3_validator
)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{s3_prototype}{PARAM_DESCRIPTION}

\item{s3_constructor}{PARAM_DESCRIPTION}

\item{s3_validator}{PARAM_DESCRIPTION}
}
\description{
make_ready_s3_valid_instance() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ready a S3 valid an instance.NA
}
\keyword{internal}
