% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ready_s3_validator}
\alias{make_ready_s3_validator}
\title{Make ready S3 validator}
\usage{
make_ready_s3_validator(
  type,
  class_name,
  s3_prototype,
  min_max_values,
  start_end_values,
  values
)
}
\arguments{
\item{type}{PARAM_DESCRIPTION}

\item{class_name}{PARAM_DESCRIPTION}

\item{s3_prototype}{PARAM_DESCRIPTION}

\item{min_max_values}{PARAM_DESCRIPTION}

\item{start_end_values}{PARAM_DESCRIPTION}

\item{values}{PARAM_DESCRIPTION}
}
\description{
make_ready_s3_validator() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ready a S3 validator. The function is called for its side effects and does not return a value.
}
\keyword{internal}
