% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ready_s4}
\alias{make_ready_s4}
\title{Make ready}
\usage{
make_ready_s4(
  name_stub,
  name_prefix = "ready4_",
  output_folder = "data-raw",
  output_sub_folder = NULL,
  class_desc = "",
  parent = NULL,
  class_slots,
  type,
  meaningful_names = NULL,
  values = NULL,
  allowed_values = NULL,
  include_classes = NULL,
  prototype_lup,
  ignore_ns_chr = NA_character_,
  required_pckg_chr_vec = NA_character_,
  names_include = NULL,
  not_same_length = NULL,
  print_set_class = TRUE,
  print_helper = TRUE,
  print_accessors = TRUE,
  print_validator = TRUE,
  print_meaningful_names = TRUE,
  class_in_cache_logic_chr = "stop"
)
}
\arguments{
\item{name_stub}{PARAM_DESCRIPTION}

\item{name_prefix}{PARAM_DESCRIPTION, Default: 'ready4_'}

\item{output_folder}{PARAM_DESCRIPTION, Default: 'data-raw'}

\item{output_sub_folder}{PARAM_DESCRIPTION, Default: NULL}

\item{class_desc}{PARAM_DESCRIPTION, Default: ''}

\item{parent}{PARAM_DESCRIPTION, Default: NULL}

\item{class_slots}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}

\item{meaningful_names}{PARAM_DESCRIPTION, Default: NULL}

\item{values}{PARAM_DESCRIPTION, Default: NULL}

\item{allowed_values}{PARAM_DESCRIPTION, Default: NULL}

\item{include_classes}{PARAM_DESCRIPTION, Default: NULL}

\item{prototype_lup}{Prototype (a lookup table)}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1), Default: 'NA'}

\item{required_pckg_chr_vec}{Required pckg (a character vector), Default: 'NA'}

\item{names_include}{PARAM_DESCRIPTION, Default: NULL}

\item{not_same_length}{PARAM_DESCRIPTION, Default: NULL}

\item{print_set_class}{PARAM_DESCRIPTION, Default: TRUE}

\item{print_helper}{PARAM_DESCRIPTION, Default: TRUE}

\item{print_accessors}{PARAM_DESCRIPTION, Default: TRUE}

\item{print_validator}{PARAM_DESCRIPTION, Default: TRUE}

\item{print_meaningful_names}{PARAM_DESCRIPTION, Default: TRUE}

\item{class_in_cache_logic_chr}{Class in cache logic (a character vector of length 1), Default: 'stop'}
}
\description{
make_ready_s4() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ready. The function is called for its side effects and does not return a value.
}
\keyword{internal}
