% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_s3_components_ls}
\alias{make_s3_components_ls}
\title{Make S3 components}
\usage{
make_s3_components_ls(
  class_name_chr,
  type_chr,
  type_ns_chr,
  type_checker_pfx_chr,
  values_chr,
  ordered_lgl,
  parent_chr,
  prototype_lup,
  min_max_values_dbl_vec,
  start_end_values_dbl_vec,
  ignore_ns_chr
)
}
\arguments{
\item{class_name_chr}{Class name (a character vector of length 1)}

\item{type_chr}{Type (a character vector of length 1)}

\item{type_ns_chr}{Type namespace (a character vector of length 1)}

\item{type_checker_pfx_chr}{Type checker prefix (a character vector of length 1)}

\item{values_chr}{Values (a character vector of length 1)}

\item{ordered_lgl}{Ordered (a logical vector of length 1)}

\item{parent_chr}{Parent (a character vector of length 1)}

\item{prototype_lup}{Prototype (a lookup table)}

\item{min_max_values_dbl_vec}{Min max values (a double vector)}

\item{start_end_values_dbl_vec}{Start end values (a double vector)}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1)}
}
\description{
make_s3_components_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a S3 components. The function is called for its side effects and does not return a value.
}
\keyword{internal}
