% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_s4_mk_tb_r3}
\alias{make_s4_mk_tb_r3}
\title{Make S4 mk}
\usage{
make_s4_mk_tb_r3(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
make_s4_mk_tb_r3() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a S4 mk.NA
}
\keyword{internal}
