% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_s4_mk_tb_row}
\alias{make_s4_mk_tb_row}
\title{Make S4 mk tibble row}
\usage{
make_s4_mk_tb_row(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
make_s4_mk_tb_row() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a S4 mk a tibble row. The function is called for its side effects and does not return a value.
}
\keyword{internal}
