% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_s4_mthds_ls}
\alias{make_s4_mthds_ls}
\title{make S4 methods list}
\usage{
make_s4_mthds_ls(fns_dir_1L_chr = "data-raw/s4_fns")
}
\arguments{
\item{fns_dir_1L_chr}{Functions directory (a character vector of length one), Default: 'data-raw/s4_fns'}
}
\value{
a S4 methods (a list)
}
\description{
make_s4_mthds_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make s4 methods list. The function returns a S4 methods (a list).
}
\keyword{internal}
