% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_show_mthd_fn}
\alias{make_show_mthd_fn}
\title{Make show method}
\usage{
make_show_mthd_fn(class_name, meaningful_names)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{meaningful_names}{PARAM_DESCRIPTION}
}
\description{
make_show_mthd_fn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make show mthd.The function is called for its side effects and does not return a value.
}
