% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_make_tb.R
\name{new_ready4_class_make_tb}
\alias{new_ready4_class_make_tb}
\title{NA}
\usage{
new_ready4_class_make_tb(x)
}
\arguments{
\item{x}{A prototype for the ready4_class_make_tb class}
}
\value{
An unvalidated instance of the ready4_class_make_tb class
}
\description{
Create a new unvalidated instance of the S3 class: ready4_class_make_tb
}
\details{
MAKE CLASS TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
