% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{new_ready4_class_pt_lup}
\alias{new_ready4_class_pt_lup}
\title{NA}
\usage{
new_ready4_class_pt_lup(x)
}
\arguments{
\item{x}{A prototype for the ready4_class_pt_lup class}
}
\value{
An unvalidated instance of the ready4_class_pt_lup class
}
\description{
Create a new unvalidated instance of the S3 class: ready4_class_pt_lup
}
\details{
PROTOTYPE LOOKUP Readyforwhatsnext S3 class of metadata of prototype classes.
}
