% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{order_tb}
\alias{order_tb}
\title{Order tibble}
\usage{
order_tb(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\description{
order_tb() is an Order Tibble generic that orders a tibble.NA
}
