% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_order_tb.R
\name{order_tb.ready4_class_make_tb}
\alias{order_tb.ready4_class_make_tb}
\title{Order tibble method applied to Class Make Table readyforwhatsnext S3 class.}
\usage{
\method{order_tb}{ready4_class_make_tb}(x, name_prefix)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{name_prefix}{PARAM_DESCRIPTION}
}
\description{
order_tb.ready4_class_make_tb() is an Order Tibble method that orders a tibble. This method is implemented for the Class Make Table readyforwhatsnext S3 class.NA
}
