% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_order_tb.R
\name{order_tb.ready4_constructor_tbl}
\alias{order_tb.ready4_constructor_tbl}
\title{Order tibble method applied to readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE.}
\usage{
\method{order_tb}{ready4_constructor_tbl}(x, name_pfx_1L_chr)
}
\arguments{
\item{x}{An instance of readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one)}
}
\value{
Instance (a readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE)
}
\description{
order_tb.ready4_constructor_tbl() is an Order Tibble method that orders a tibble. This method is implemented for the readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE.NA
}
