% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{ready4_class_pt_lup}
\alias{ready4_class_pt_lup}
\title{NA}
\usage{
ready4_class_pt_lup(x = make_prototype_ready4_class_pt_lup())
}
\arguments{
\item{x}{A prototype for the ready4_class_pt_lup class, Default: make_prototype_ready4_class_pt_lup()}
}
\value{
A validated instance of the ready4_class_pt_lup class
}
\description{
Create a new valid instance of the S3 class: ready4_class_pt_lup
}
\details{
PROTOTYPE LOOKUP Readyforwhatsnext S3 class of metadata of prototype classes.
}
