% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_constructor_tbl.R
\name{ready4_constructor_tbl}
\alias{ready4_constructor_tbl}
\title{readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE}
\usage{
ready4_constructor_tbl(x = make_pt_ready4_constructor_tbl())
}
\arguments{
\item{x}{A prototype for the readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE, Default: make_pt_ready4_constructor_tbl()}
}
\value{
A validated instance of the readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE
}
\description{
Create a new valid instance of the readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE
}
\details{
Readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE of metadata required to make new classes.
}
