% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4class.R
\docType{package}
\name{ready4class-package}
\alias{ready4class}
\alias{ready4class-package}
\title{ready4class Class creation tools}
\description{
ready4class is a package of classes and methods to help standardise and streamline the process for creating and extending S3 and S3 classes and
methods for use with the readyforwhatsnext framework for data synthesis and simulation in mental health.
}
\details{
To learn more about ready4use, start with the vignettes:
`browseVignettes(package = "ready4class")`
}
\section{First section}{


Some text.
}

\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
