% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor.R
\name{ready4class_constructor}
\alias{ready4class_constructor}
\title{ready4 S3 class Constructor Table.}
\usage{
ready4class_constructor(x = make_pt_ready4class_constructor())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class Constructor Table., Default: make_pt_ready4class_constructor()}
}
\value{
A validated instance of the ready4 S3 class Constructor Table.
}
\description{
Create a new valid instance of the ready4 S3 class Constructor Table.
}
\details{
ready4 S3 class Constructor Table.
}
