% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor_tbl.R
\name{ready4class_constructor_tbl}
\alias{ready4class_constructor_tbl}
\title{ready4 S3 class Constructor Table of metadata required to make new classes.}
\usage{
ready4class_constructor_tbl(x = make_pt_ready4class_constructor_tbl())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class Constructor Table of metadata required to make new classes., Default: make_pt_ready4class_constructor_tbl()}
}
\value{
A validated instance of the ready4 S3 class Constructor Table of metadata required to make new classes.
}
\description{
Create a new valid instance of the ready4 S3 class Constructor Table of metadata required to make new classes.
}
\details{
ready4 S3 class Constructor Table of metadata required to make new classes.
}
