% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_pt_lup.R
\name{make_new_ready4class_pt_lup}
\alias{make_new_ready4class_pt_lup}
\alias{make_pt_ready4class_pt_lup}
\alias{validate_ready4class_pt_lup}
\alias{is_ready4class_pt_lup}
\title{make new ready4class package prototype lookup table ready4 S3 class Prototype Lookup Table of class metadata.}
\usage{
make_new_ready4class_pt_lup(x)

make_pt_ready4class_pt_lup(
  type_chr = character(0),
  val_chr = character(0),
  pt_ns_chr = character(0),
  fn_to_call_chr = character(0),
  default_val_chr = character(0),
  old_class_lgl = logical(0)
)

validate_ready4class_pt_lup(x)

is_ready4class_pt_lup(x)
}
\arguments{
\item{x}{An object of any type}

\item{type_chr}{Type (a character vector), Default: character(0)}

\item{val_chr}{Value (a character vector), Default: character(0)}

\item{pt_ns_chr}{Prototype namespace (a character vector), Default: character(0)}

\item{fn_to_call_chr}{Function to call (a character vector), Default: character(0)}

\item{default_val_chr}{Default value (a character vector), Default: character(0)}

\item{old_class_lgl}{Old class (a logical vector), Default: logical(0)}
}
\value{
An unvalidated instance of the ready4 S3 class Prototype Lookup Table of class metadata.

A prototype for ready4 S3 class Prototype Lookup Table of class metadata.

A prototpe for ready4 S3 class Prototype Lookup Table of class metadata.

A logical value, TRUE if a valid instance of the ready4 S3 class Prototype Lookup Table of class metadata.
}
\description{
make new ready4class package prototype lookup table ready4 S3 class Prototype Lookup Table of class metadata.

make prototype ready4class package prototype lookup table ready4 S3 class Prototype Lookup Table of class metadata.

validate ready4class package prototype lookup table ready4 S3 class Prototype Lookup Table of class metadata.

is ready4class package prototype lookup table ready4 S3 class Prototype Lookup Table of class metadata.
}
\details{
ready4 S3 class Prototype Lookup Table.

ready4 S3 class Prototype Lookup Table.

ready4 S3 class Prototype Lookup Table.

ready4 S3 class Prototype Lookup Table.
}
\section{Functions}{
\itemize{
\item \code{make_new_ready4class_pt_lup}: Create a new unvalidated instance of the ready4 S3 class Prototype Lookup Table of class metadata.

\item \code{make_pt_ready4class_pt_lup}: Create a new prototype for the ready4 S3 class Prototype Lookup Table of class metadata.

\item \code{validate_ready4class_pt_lup}: Validate an instance of the ready4 S3 class Prototype Lookup Table of class metadata.

\item \code{is_ready4class_pt_lup}: Check whether an object is a valid instance of the ready4 S3 class Prototype Lookup Table of class metadata.
}}

