% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_pt_lup.R
\name{ready4class_pt_lup}
\alias{ready4class_pt_lup}
\title{ready4 S3 class Prototype Lookup Table of class metadata.}
\usage{
ready4class_pt_lup(x = make_pt_ready4class_pt_lup())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class Prototype Lookup Table of class metadata., Default: make_pt_ready4class_pt_lup()}
}
\value{
A validated instance of the ready4 S3 class Prototype Lookup Table of class metadata.
}
\description{
Create a new valid instance of the ready4 S3 class Prototype Lookup Table of class metadata.
}
\details{
ready4 S3 class Prototype Lookup Table.
}
