% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_remake_ls_cols.R
\name{remake_ls_cols}
\alias{remake_ls_cols}
\alias{remake_ls_cols.ready4_constructor_tbl}
\alias{remake_ls_cols,ready4_constructor_tbl-method}
\title{Remake list columns}
\usage{
remake_ls_cols(x, ...)

remake_ls_cols.ready4_constructor_tbl(x)

\S4method{remake_ls_cols}{ready4_constructor_tbl}(x)
}
\arguments{
\item{x}{An instance of readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE}

\item{...}{Additional arguments}
}
\value{
Instance (a readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE)
}
\description{
remake_ls_cols() is a Remake List Columns generic that remakes list columns.

remake_ls_cols.ready4_constructor_tbl() is a Remake List Columns method that remakes list columns. This method is implemented for the readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE. The function is called for its side effects and does not return a value.
}
