% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{remake_ls_cols}
\alias{remake_ls_cols}
\title{Remake list columns}
\usage{
remake_ls_cols(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{PARAM_DESCRIPTION}
}
\description{
remake_ls_cols() is a Remake List Columns generic that remakes list columns.NA
}
