% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_remake_ls_cols.R
\name{remake_ls_cols.ready4_class_make_tb}
\alias{remake_ls_cols.ready4_class_make_tb}
\title{Remake list columns method applied to Class Make Table readyforwhatsnext S3 class.}
\usage{
\method{remake_ls_cols}{ready4_class_make_tb}(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
remake_ls_cols.ready4_class_make_tb() is a Remake List Columns method that remakes list columns. This method is implemented for the Class Make Table readyforwhatsnext S3 class.NA
}
