% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_remake_ls_cols.R
\name{remake_ls_cols.ready4_constructor_tbl}
\alias{remake_ls_cols.ready4_constructor_tbl}
\title{Remake list columns method applied to CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class.}
\usage{
\method{remake_ls_cols}{ready4_constructor_tbl}(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
remake_ls_cols.ready4_constructor_tbl() is a Remake List Columns method that remakes list columns. This method is implemented for the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class.NA
}
\keyword{internal}
