% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_class_from_cache}
\alias{remove_class_from_cache}
\title{Remove class from cache}
\usage{
remove_class_from_cache(class_nm_1L_chr, class_in_cache_cdn_1L_chr = "stop")
}
\arguments{
\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{class_in_cache_cdn_1L_chr}{Class in cache condition (a character vector of length one), Default: 'stop'}
}
\value{
No return value, called for side effects.
}
\description{
remove_class_from_cache() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove class from cache. The function is called for its side effects and does not return a value.
}
\keyword{internal}
