% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_replace.R
\name{replace_NA_in_fn}
\alias{replace_NA_in_fn}
\title{Replace NA in}
\usage{
replace_NA_in_fn(fn_body_1L_chr)
}
\arguments{
\item{fn_body_1L_chr}{Function body (a character vector of length one)}
}
\value{
Function body (a character vector of length one)
}
\description{
replace_NA_in_fn() is a Replace function that edits an object, replacing a specified element with another specified element. Specifically, this function implements an algorithm to replace na in function. Function argument fn_body_1L_chr specifies the object to be updated.The function returns a function body (a character vector of length one).
}
