% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_resolve.R
\name{resolve_parent_ns_chr}
\alias{resolve_parent_ns_chr}
\title{Resolve parent namespace}
\usage{
resolve_parent_ns_chr(parent_ns_ls)
}
\arguments{
\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
resolve_parent_ns_chr() is a Resolve function that resolves inconsistencies between two or more objects. Specifically, this function implements an algorithm to resolve parent a namespace. The function is called for its side effects and does not return a value.
}
\keyword{internal}
