% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_set.R
\name{set_old_classes_extend_tb}
\alias{set_old_classes_extend_tb}
\title{Set old classes extend}
\usage{
set_old_classes_extend_tb(type, prototype_lup)
}
\arguments{
\item{type}{PARAM_DESCRIPTION}

\item{prototype_lup}{Prototype (a lookup table)}
}
\description{
set_old_classes_extend_tb() is a Set function that sets the value of an object. Specifically, this function implements an algorithm to set old classes extend. The function is called for its side effects and does not return a value.
}
\keyword{internal}
