% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_set.R
\name{set_ready_class}
\alias{set_ready_class}
\title{Set ready class}
\usage{
set_ready_class(
  class_name,
  class_slots,
  type,
  proto_ls,
  parent,
  print_set_class,
  class_desc,
  output_file_class,
  include_classes,
  prototype_lup,
  helper_lgl = F,
  parent_ns_ls
)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{class_slots}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}

\item{proto_ls}{Proto (a list)}

\item{parent}{PARAM_DESCRIPTION}

\item{print_set_class}{PARAM_DESCRIPTION}

\item{class_desc}{PARAM_DESCRIPTION}

\item{output_file_class}{PARAM_DESCRIPTION}

\item{include_classes}{PARAM_DESCRIPTION}

\item{prototype_lup}{Prototype (a lookup table)}

\item{helper_lgl}{Helper (a logical vector of length 1), Default: F}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
set_ready_class() is a Set function that sets the value of an object. Specifically, this function implements an algorithm to set ready class.NA
}
\keyword{internal}
