% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_simplify.R
\name{simplify_class_name}
\alias{simplify_class_name}
\title{Simplify class name}
\usage{
simplify_class_name(st_class_fn_chr, package_chr)
}
\arguments{
\item{st_class_fn_chr}{St class function (a character vector of length 1)}

\item{package_chr}{Package (a character vector of length 1)}
}
\description{
simplify_class_name() is a Simplify function that simplifies and object. Specifically, this function implements an algorithm to simplify class a name.NA
}
\keyword{internal}
