% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_alg_to_ref_cls_nm}
\alias{transform_alg_to_ref_cls_nm}
\title{Transform algorithm to reference class name}
\usage{
transform_alg_to_ref_cls_nm(st_class_fn_chr, package_chr)
}
\arguments{
\item{st_class_fn_chr}{St class function (a character vector)}

\item{package_chr}{Package (a character vector)}
}
\description{
transform_alg_to_ref_cls_nm() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to a transform alg to ref cls a name. Function argument st_class_fn_chr specifies the object to be updated. Argument package_chr provides the object to be updated.The function is called for its side effects and does not return a value.
}
