% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_class_ns}
\alias{transform_class_ns}
\title{transform class namespace}
\usage{
transform_class_ns(class_ns_1L_chr, dev_pkg_ns_1L_chr)
}
\arguments{
\item{class_ns_1L_chr}{Class namespace (a character vector of length one)}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one)}
}
\value{
Transformed class namespace (a character vector of length one)
}
\description{
transform_class_ns() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform class namespace. Function argument class_ns_1L_chr specifies the object to be updated. Argument dev_pkg_ns_1L_chr provides the object to be updated. The function returns Transformed class namespace (a character vector of length one).
}
\keyword{internal}
