% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_fn_into_chr}
\alias{transform_fn_into_chr}
\title{Transform function into}
\usage{
transform_fn_into_chr(fn)
}
\arguments{
\item{fn}{Function (a function)}
}
\value{
Function (a character vector of length one)
}
\description{
transform_fn_into_chr() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform function into character vector. Function argument fn specifies the object to be updated. The function returns Function (a character vector of length one).
}
