% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_new_classes_ls}
\alias{transform_new_classes_ls}
\title{Transform new classes}
\usage{
transform_new_classes_ls(new_classes_ls)
}
\arguments{
\item{new_classes_ls}{New classes (a list)}
}
\description{
transform_new_classes_ls() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform new classes. Function argument new_classes_ls specifies the object to be updated. Argument NA provides the object to be updated.NA
}
\keyword{internal}
