% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_parent_ns_ls}
\alias{transform_parent_ns_ls}
\title{Transform parent namespace list}
\usage{
transform_parent_ns_ls(parent_ns_ls)
}
\arguments{
\item{parent_ns_ls}{Parent namespace (a list)}
}
\value{
Transformed parent namespace (a list)
}
\description{
transform_parent_ns_ls() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform parent namespace list. Function argument parent_ns_ls specifies the object to be updated. The function returns Transformed parent namespace (a list).
}
