% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_pt_ls_for_new_clss}
\alias{transform_pt_ls_for_new_clss}
\title{Transform prototype list for new classes}
\usage{
transform_pt_ls_for_new_clss(pts_for_new_clss_ls)
}
\arguments{
\item{pts_for_new_clss_ls}{Prototypes for new classes (a list)}
}
\value{
Transformed prototypes for new classes (a list)
}
\description{
transform_pt_ls_for_new_clss() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform prototype list for new classes. The function returns Transformed prototypes for new classes (a list).
}
\keyword{internal}
