% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{update_lup_for_ns}
\alias{update_lup_for_ns}
\title{Update lookup table for namespace}
\usage{
update_lup_for_ns(x, attached_nss_chr, ...)
}
\arguments{
\item{x}{An object}

\item{attached_nss_chr}{Attached namespaces (a character vector)}

\item{...}{PARAM_DESCRIPTION}
}
\description{
update_lup_for_ns() is an Update Lookup Table for Namespace generic that updates a lookup table with namespace data.NA
}
