% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_update_lup_for_ns.R
\name{update_lup_for_ns.ready4_class_pt_lup}
\alias{update_lup_for_ns.ready4_class_pt_lup}
\title{Update lookup table for namespace method applied to Class Prototype Lookup Table readyforwhatsnext S3 class.}
\usage{
\method{update_lup_for_ns}{ready4_class_pt_lup}(x, namespace_contexts)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{namespace_contexts}{PARAM_DESCRIPTION}
}
\description{
update_lup_for_ns.ready4_class_pt_lup() is an Update Lookup Table for Namespace method that updates a lookup table with namespace data. This method is implemented for the Class Prototype Lookup Table readyforwhatsnext S3 class.NA
}
\keyword{internal}
