% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_validate.R
\name{validate_ready}
\alias{validate_ready}
\title{Validate ready}
\usage{
validate_ready(
  class_name,
  parent,
  not_same_length = NULL,
  allowed_values = NULL,
  names_include = NULL,
  print_validator = FALSE
)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION}

\item{not_same_length}{PARAM_DESCRIPTION, Default: NULL}

\item{allowed_values}{PARAM_DESCRIPTION, Default: NULL}

\item{names_include}{PARAM_DESCRIPTION, Default: NULL}

\item{print_validator}{PARAM_DESCRIPTION, Default: FALSE}
}
\description{
validate_ready() is a Validate function that validates an object. Specifically, this function implements an algorithm to validate ready. The function is called for its side effects and does not return a value.
}
\keyword{internal}
