% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_make_tb.R
\name{validate_ready4_class_make_tb}
\alias{validate_ready4_class_make_tb}
\title{Class Make Table readyforwhatsnext S3 class}
\usage{
validate_ready4_class_make_tb(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4_class_make_tb class}
}
\value{
A prototpe for
}
\description{
Validate an instance of the
}
\details{
MAKE CLASS TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
