% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{validate_ready4_class_pt_lup}
\alias{validate_ready4_class_pt_lup}
\title{NA}
\usage{
validate_ready4_class_pt_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4_class_pt_lup class}
}
\value{
A prototpe for ready4_class_pt_lup class
}
\description{
Validate an instance of the S3 class: ready4_class_pt_lup
}
\details{
PROTOTYPE LOOKUP Readyforwhatsnext S3 class of metadata of prototype classes.
}
