% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_constructor_tbl.R
\name{validate_ready4_constructor_tbl}
\alias{validate_ready4_constructor_tbl}
\title{Validate CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class}
\usage{
validate_ready4_constructor_tbl(x)
}
\arguments{
\item{x}{An unvalidated instance of the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class}
}
\value{
A prototpe for CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\description{
Validate an instance of the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\details{
CLASS CONSTRUCTOR TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
