% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor.R
\name{validate_ready4class_constructor}
\alias{validate_ready4class_constructor}
\title{validate ready4class constructor Class constructor table}
\usage{
validate_ready4class_constructor(x)
}
\arguments{
\item{x}{An unvalidated instance of the Class constructor table}
}
\value{
A prototpe for Class constructor table
}
\description{
Validate an instance of the Class constructor table
}
\details{
Class constructor table
}
\keyword{internal}
