% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor_tbl.R
\name{validate_ready4class_constructor_tbl}
\alias{validate_ready4class_constructor_tbl}
\title{Validate ready4class package constructor table ready4 S3 class Constructor Table of metadata required to make new classes.}
\usage{
validate_ready4class_constructor_tbl(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class Constructor Table of metadata required to make new classes.}
}
\value{
A prototpe for ready4 S3 class Constructor Table of metadata required to make new classes.
}
\description{
Validate an instance of the ready4 S3 class Constructor Table of metadata required to make new classes.
}
\details{
ready4 S3 class Constructor Table.
}
