% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_manifest.R
\name{validate_ready4class_manifest}
\alias{validate_ready4class_manifest}
\title{validate ready4class manifest Class creation manifest}
\usage{
validate_ready4class_manifest(x)
}
\arguments{
\item{x}{An unvalidated instance of the Class creation manifest}
}
\value{
A prototpe for Class creation manifest
}
\description{
Validate an instance of the Class creation manifest
}
\details{
Class creation manifest
}
\keyword{internal}
