% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_pt_lup.R
\name{validate_ready4class_pt_lup}
\alias{validate_ready4class_pt_lup}
\title{validate ready4class prototype lookup table Class prototype lookup table}
\usage{
validate_ready4class_pt_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the Class prototype lookup table}
}
\value{
A prototpe for Class prototype lookup table
}
\description{
Validate an instance of the Class prototype lookup table
}
\details{
Class prototype lookup table
}
\keyword{internal}
