% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_accessors}
\alias{write_accessors}
\title{Write accessors}
\usage{
write_accessors(
  slot_name_chr,
  set_only,
  import_packages_ls,
  class_name,
  print_accessors,
  output_folder
)
}
\arguments{
\item{slot_name_chr}{Slot name (a character vector of length 1)}

\item{set_only}{PARAM_DESCRIPTION}

\item{import_packages_ls}{Import packages (a list)}

\item{class_name}{PARAM_DESCRIPTION}

\item{print_accessors}{PARAM_DESCRIPTION}

\item{output_folder}{PARAM_DESCRIPTION}
}
\description{
write_accessors() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write accessors.NA
}
\keyword{internal}
