% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_write_classes.R
\name{write_classes}
\alias{write_classes}
\alias{write_classes.ready4_constructor_tbl}
\alias{write_classes,ready4_constructor_tbl-method}
\title{Write classes}
\usage{
write_classes(x, ...)

write_classes.ready4_constructor_tbl(
  x,
  name_pfx_1L_chr,
  output_dir_1L_chr,
  file_exists_cdn_1L_chr = NULL,
  prototype_lup = NULL,
  dev_pkg_ns_1L_chr = ready4fun::get_dev_pkg_nm(),
  nss_to_ignore_chr,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup = NULL
)

\S4method{write_classes}{ready4_constructor_tbl}(
  x,
  name_pfx_1L_chr,
  output_dir_1L_chr,
  file_exists_cdn_1L_chr = NULL,
  prototype_lup = NULL,
  dev_pkg_ns_1L_chr = ready4fun::get_dev_pkg_nm(),
  nss_to_ignore_chr,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup = NULL
)
}
\arguments{
\item{x}{An instance of ready4 S3 class CLASS CONSTRUCTOR TABLE}

\item{...}{Additional arguments}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one)}

\item{file_exists_cdn_1L_chr}{File exists condition (a character vector of length one), Default: NULL}

\item{prototype_lup}{Prototype (a lookup table), Default: NULL}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one), Default: ready4fun::get_dev_pkg_nm()}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector)}

\item{req_pkgs_chr}{Req packages (a character vector), Default: 'NA'}

\item{class_in_cache_cdn_1L_chr}{Class in cache condition (a character vector of length one), Default: 'stop'}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\description{
write_classes() is a Write Classes generic that writes new classes.

write_classes.ready4_constructor_tbl() is a Write Classes method that writes new classes. This method is implemented for the ready4 S3 class CLASS CONSTRUCTOR TABLE. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
