% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_write_classes.R
\name{write_classes.ready4_constructor_tbl}
\alias{write_classes.ready4_constructor_tbl}
\title{Write classes method applied to CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class.}
\usage{
\method{write_classes}{ready4_constructor_tbl}(
  x,
  name_pfx_1L_chr,
  output_dir_1L_chr,
  file_exists_cdn_1L_chr = NULL,
  prototype_lup = NULL,
  nss_to_ignore_chr,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup = NULL
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one)}

\item{file_exists_cdn_1L_chr}{File exists condition (a character vector of length one), Default: NULL}

\item{prototype_lup}{Prototype (a lookup table), Default: NULL}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector)}

\item{req_pkgs_chr}{Req packages (a character vector), Default: 'NA'}

\item{class_in_cache_cdn_1L_chr}{Class in cache condition (a character vector of length one), Default: 'stop'}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\description{
write_classes.ready4_constructor_tbl() is a Write Classes method that writes new classes. This method is implemented for the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class.NA
}
