% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{write_classes_and_make_lup}
\alias{write_classes_and_make_lup}
\title{Write classes and make lookup table}
\usage{
write_classes_and_make_lup(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{PARAM_DESCRIPTION}
}
\description{
write_classes_and_make_lup() is a Write Classes and Make Lookup Table generic that makes new classes and creates or updates a class prototype lookup table.NA
}
