% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_fn_txt_and_tags}
\alias{write_fn_txt_and_tags}
\title{Write function txt and tags}
\usage{
write_fn_txt_and_tags(fn_name, fn_text, fn_type, class_name, class_desc)
}
\arguments{
\item{fn_name}{PARAM_DESCRIPTION}

\item{fn_text}{PARAM_DESCRIPTION}

\item{fn_type}{PARAM_DESCRIPTION}

\item{class_name}{PARAM_DESCRIPTION}

\item{class_desc}{PARAM_DESCRIPTION}
}
\description{
write_fn_txt_and_tags() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a function txt and tags.NA
}
\keyword{internal}
