% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_gen_meth}
\alias{write_gen_meth}
\title{Write gen meth}
\usage{
write_gen_meth(
  fn_name_chr,
  args_chr_vec = c("x"),
  signature_chr = NA_character_,
  package_chr = NA_character_,
  where_chr = NA_character_,
  class_chr,
  fn,
  fn_type_chr_vec,
  fn_desc_chr_vec = rep(NA_character_, 2),
  fn_title_chr = NA_character_,
  fn_out_type_chr = NA_character_,
  import_chr_vec,
  write_file_ls,
  output_dir_chr,
  append_lgl = T,
  doc_in_class_lgl = F,
  generic_exists_lgl,
  overwrite_lgl = F,
  s3_lgl,
  write_lgl
)
}
\arguments{
\item{fn_name_chr}{Function name (a character vector of length 1)}

\item{args_chr_vec}{Arguments (a character vector), Default: c("x")}

\item{signature_chr}{Signature (a character vector of length 1), Default: 'NA'}

\item{package_chr}{Package (a character vector of length 1), Default: 'NA'}

\item{where_chr}{Where (a character vector of length 1), Default: 'NA'}

\item{class_chr}{Class (a character vector of length 1)}

\item{fn}{Function (a function)}

\item{fn_type_chr_vec}{Function type (a character vector)}

\item{fn_desc_chr_vec}{Function description (a character vector), Default: rep(NA_character_, 2)}

\item{fn_title_chr}{Function title (a character vector of length 1), Default: 'NA'}

\item{fn_out_type_chr}{Function out type (a character vector of length 1), Default: 'NA'}

\item{import_chr_vec}{Import (a character vector)}

\item{write_file_ls}{Write file (a list)}

\item{output_dir_chr}{Output directory (a character vector of length 1)}

\item{append_lgl}{Append (a logical vector of length 1), Default: T}

\item{doc_in_class_lgl}{Document in class (a logical vector of length 1), Default: F}

\item{generic_exists_lgl}{Generic exists (a logical vector of length 1)}

\item{overwrite_lgl}{Overwrite (a logical vector of length 1), Default: F}

\item{s3_lgl}{S3 (a logical vector of length 1)}

\item{write_lgl}{Write (a logical vector of length 1)}
}
\description{
write_gen_meth() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write gen meth. The function is called for its side effects and does not return a value.
}
\keyword{internal}
