% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_gtr_str_mthds_for_slots}
\alias{write_gtr_str_mthds_for_slots}
\title{Write getter setter methods for slots}
\usage{
write_gtr_str_mthds_for_slots(
  slot_names_chr_vec,
  set_only,
  parent,
  class_name,
  print_accessors,
  output_folder,
  ignore_ns_chr,
  required_pckg_chr_vec
)
}
\arguments{
\item{slot_names_chr_vec}{PARAM_DESCRIPTION}

\item{set_only}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION}

\item{class_name}{PARAM_DESCRIPTION}

\item{print_accessors}{PARAM_DESCRIPTION}

\item{output_folder}{PARAM_DESCRIPTION}

\item{ignore_ns_chr}{Ignore namespace (a character vector)}

\item{required_pckg_chr_vec}{PARAM_DESCRIPTION}
}
\description{
write_gtr_str_mthds_for_slots() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a getter setter mthds for slots.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
