% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_r4_mthds}
\alias{write_r4_mthds}
\title{Write ready4 S4 methods}
\usage{
write_r4_mthds(
  s4_mthds_ls,
  pkg_nm_1L_chr = character(0),
  output_dir_1L_chr = "R",
  import_from_chr = character(0)
)
}
\arguments{
\item{s4_mthds_ls}{S4 methods (a list)}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: character(0)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one), Default: 'R'}

\item{import_from_chr}{Import from (a character vector), Default: character(0)}
}
\value{
Writen files (a list of lists)
}
\description{
write_r4_mthds() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write ready4 s4 methods. The function returns Writen files (a list of lists).
}
\keyword{internal}
