% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_s3_s4}
\alias{write_s3_s4}
\title{Write S3}
\usage{
write_s3_s4(methods_tb, fn_ls, package_chr, output_dir_chr)
}
\arguments{
\item{methods_tb}{Methods (a tibble)}

\item{fn_ls}{Function list (a list of functions)}

\item{package_chr}{Package (a character vector of length 1)}

\item{output_dir_chr}{Output directory (a character vector of length 1)}
}
\description{
write_s3_s4() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a S3. The function is called for its side effects and does not return a value.
}
\keyword{internal}
