% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_script_to_make_gnrc}
\alias{write_script_to_make_gnrc}
\title{write script to make generic}
\usage{
write_script_to_make_gnrc(
  write_file_ls,
  gnrc_exists_1L_lgl,
  gen_mthd_pair_ls,
  fn_name_1L_chr,
  fn_type_1L_chr,
  fn_desc_1L_chr = NA_character_,
  fn_outp_type_1L_chr = NA_character_,
  fn_title_1L_chr = NA_character_,
  class_nm_1L_chr = NA_character_,
  output_dir_1L_chr = NA_character_,
  overwrite_1L_lgl = F,
  s3_1L_lgl = F,
  write_1L_lgl = T,
  doc_in_class_1L_lgl = F,
  fn_types_lup = NULL,
  object_type_lup = NULL,
  consent_1L_chr = NULL
)
}
\arguments{
\item{write_file_ls}{Write file (a list)}

\item{gnrc_exists_1L_lgl}{Generic exists (a logical vector of length one)}

\item{gen_mthd_pair_ls}{Generate method pair (a list)}

\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{fn_type_1L_chr}{Function type (a character vector of length one)}

\item{fn_desc_1L_chr}{Function description (a character vector of length one), Default: 'NA'}

\item{fn_outp_type_1L_chr}{Function output type (a character vector of length one), Default: 'NA'}

\item{fn_title_1L_chr}{Function title (a character vector of length one), Default: 'NA'}

\item{class_nm_1L_chr}{Class name (a character vector of length one), Default: 'NA'}

\item{output_dir_1L_chr}{Output directory (a character vector of length one), Default: 'NA'}

\item{overwrite_1L_lgl}{Overwrite (a logical vector of length one), Default: F}

\item{s3_1L_lgl}{S3 (a logical vector of length one), Default: F}

\item{write_1L_lgl}{Write (a logical vector of length one), Default: T}

\item{doc_in_class_1L_lgl}{Document in class (a logical vector of length one), Default: F}

\item{fn_types_lup}{Function types (a lookup table), Default: NULL}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: NULL}
}
\description{
write_script_to_make_gnrc() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write script to make generic. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
