% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_script_to_make_mthd}
\alias{write_script_to_make_mthd}
\title{Write script to make method}
\usage{
write_script_to_make_mthd(
  write_file_ls,
  gen_mthd_pair_ls,
  class_name_chr,
  fn_name_chr,
  fn_type_chr,
  fn_desc_chr = NA_character_,
  fn_out_type_chr = NA_character_,
  import_chr_vec,
  write_lgl = T,
  append_lgl = T,
  doc_in_class_lgl = F
)
}
\arguments{
\item{write_file_ls}{Write file (a list)}

\item{gen_mthd_pair_ls}{Generate method pair (a list)}

\item{class_name_chr}{Class name (a character vector)}

\item{fn_name_chr}{Function name (a character vector)}

\item{fn_type_chr}{Function type (a character vector)}

\item{fn_desc_chr}{Function description (a character vector), Default: 'NA'}

\item{fn_out_type_chr}{Function out type (a character vector), Default: 'NA'}

\item{import_chr_vec}{PARAM_DESCRIPTION}

\item{write_lgl}{Write (a logical vector), Default: T}

\item{append_lgl}{Append (a logical vector), Default: T}

\item{doc_in_class_lgl}{Document in class (a logical vector), Default: F}
}
\description{
write_script_to_make_mthd() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write script to a make mthd.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
