% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scripts_to_mk_clss}
\alias{write_scripts_to_mk_clss}
\title{Write scripts to make classes}
\usage{
write_scripts_to_mk_clss(
  new_classes_ls,
  pckg_name_chr,
  class_pfx_chr,
  R_dir_chr = "R",
  pt_lup,
  description_ls = NULL,
  ignore_ns_chr = NA_character_,
  required_pckg_chr_vec = NA_character_
)
}
\arguments{
\item{new_classes_ls}{New classes (a list)}

\item{pckg_name_chr}{Pckg name (a character vector)}

\item{class_pfx_chr}{Class prefix (a character vector)}

\item{R_dir_chr}{R directory (a character vector), Default: 'R'}

\item{pt_lup}{Prototype (a lookup table)}

\item{description_ls}{Description (a list), Default: NULL}

\item{ignore_ns_chr}{Ignore namespace (a character vector), Default: 'NA'}

\item{required_pckg_chr_vec}{PARAM_DESCRIPTION, Default: 'NA'}
}
\description{
write_scripts_to_mk_clss() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write scripts to mk clss.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
