% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scripts_to_mk_r3_cls}
\alias{write_scripts_to_mk_r3_cls}
\title{Write scripts to make ready4 S3 class}
\usage{
write_scripts_to_mk_r3_cls(
  name_stub_1L_chr,
  name_pfx_1L_chr = "ready4_",
  output_dir_1L_chr = "data-raw",
  class_desc_1L_chr = "",
  parent_cls_nm_1L_chr = NULL,
  type_1L_chr,
  pt_ns_1L_chr = "",
  pt_chkr_pfx_1L_chr = "is.",
  vals_ls = NULL,
  ordered_1L_lgl = FALSE,
  allowed_vals_ls = NULL,
  min_max_vals_dbl = NULL,
  start_end_vals_dbl = NULL,
  prototype_lup,
  dev_pkg_ns_1L_chr = ready4fun::get_dev_pkg_nm(),
  nss_to_ignore_chr = NA_character_,
  file_exists_cdn_1L_chr = "skip",
  abbreviations_lup = NULL,
  asserts_ls = NULL
)
}
\arguments{
\item{name_stub_1L_chr}{Name stub (a character vector of length one)}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one), Default: 'ready4_'}

\item{output_dir_1L_chr}{Output directory (a character vector of length one), Default: 'data-raw'}

\item{class_desc_1L_chr}{Class description (a character vector of length one), Default: ''}

\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one), Default: NULL}

\item{type_1L_chr}{Type (a character vector of length one)}

\item{pt_ns_1L_chr}{Prototype namespace (a character vector of length one), Default: ''}

\item{pt_chkr_pfx_1L_chr}{Prototype checker prefix (a character vector of length one), Default: 'is.'}

\item{vals_ls}{Values (a list), Default: NULL}

\item{ordered_1L_lgl}{Ordered (a logical vector of length one), Default: FALSE}

\item{allowed_vals_ls}{Allowed values (a list), Default: NULL}

\item{min_max_vals_dbl}{Min max values (a double vector), Default: NULL}

\item{start_end_vals_dbl}{Start end values (a double vector), Default: NULL}

\item{prototype_lup}{Prototype (a lookup table)}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one), Default: ready4fun::get_dev_pkg_nm()}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector), Default: 'NA'}

\item{file_exists_cdn_1L_chr}{File exists condition (a character vector of length one), Default: 'skip'}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}

\item{asserts_ls}{Asserts (a list), Default: NULL}
}
\description{
write_scripts_to_mk_r3_cls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write scripts to make ready4 s3 class. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
