% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scripts_to_mk_r3_clss}
\alias{write_scripts_to_mk_r3_clss}
\title{Write scripts to make readyforwhatsnext S3 classes}
\usage{
write_scripts_to_mk_r3_clss(
  name_stub,
  name_prefix = "ready4_",
  output_folder = "data-raw",
  class_desc = "",
  parent = NULL,
  type,
  type_namespace = "",
  type_checker_prefix = "is.",
  values = NULL,
  ordered = FALSE,
  allowed_values = NULL,
  min_max_values = NULL,
  start_end_values = NULL,
  prototype_lup,
  ignore_ns_chr = NA_character_,
  file_exists_logic = "skip"
)
}
\arguments{
\item{name_stub}{PARAM_DESCRIPTION}

\item{name_prefix}{PARAM_DESCRIPTION, Default: 'ready4_'}

\item{output_folder}{PARAM_DESCRIPTION, Default: 'data-raw'}

\item{class_desc}{PARAM_DESCRIPTION, Default: ''}

\item{parent}{PARAM_DESCRIPTION, Default: NULL}

\item{type}{PARAM_DESCRIPTION}

\item{type_namespace}{PARAM_DESCRIPTION, Default: ''}

\item{type_checker_prefix}{PARAM_DESCRIPTION, Default: 'is.'}

\item{values}{PARAM_DESCRIPTION, Default: NULL}

\item{ordered}{PARAM_DESCRIPTION, Default: FALSE}

\item{allowed_values}{PARAM_DESCRIPTION, Default: NULL}

\item{min_max_values}{PARAM_DESCRIPTION, Default: NULL}

\item{start_end_values}{PARAM_DESCRIPTION, Default: NULL}

\item{prototype_lup}{Prototype (a lookup table)}

\item{ignore_ns_chr}{Ignore namespace (a character vector), Default: 'NA'}

\item{file_exists_logic}{PARAM_DESCRIPTION, Default: 'skip'}
}
\description{
write_scripts_to_mk_r3_clss() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write scripts to mk a readyforwhatsnext S3 clss.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
