% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scripts_to_mk_r3_clss_checker}
\alias{write_scripts_to_mk_r3_clss_checker}
\title{Write scripts to make readyforwhatsnext S3 classes checker}
\usage{
write_scripts_to_mk_r3_clss_checker(class_name, s3_validator)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{s3_validator}{PARAM_DESCRIPTION}
}
\description{
write_scripts_to_mk_r3_clss_checker() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write scripts to mk a readyforwhatsnext S3 clss checker.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
