% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scripts_to_mk_r3_clss_pts}
\alias{write_scripts_to_mk_r3_clss_pts}
\title{Write scripts to make readyforwhatsnext S3 classes prototypes}
\usage{
write_scripts_to_mk_r3_clss_pts(
  type,
  type_namespace,
  values,
  ordered,
  class_name,
  parent_chr,
  prototype_lup
)
}
\arguments{
\item{type}{PARAM_DESCRIPTION}

\item{type_namespace}{PARAM_DESCRIPTION}

\item{values}{PARAM_DESCRIPTION}

\item{ordered}{PARAM_DESCRIPTION}

\item{class_name}{PARAM_DESCRIPTION}

\item{parent_chr}{Parent (a character vector)}

\item{prototype_lup}{Prototype (a lookup table)}
}
\description{
write_scripts_to_mk_r3_clss_pts() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write scripts to mk a readyforwhatsnext S3 clss prototypes.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
