% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scripts_to_mk_r4_cls}
\alias{write_scripts_to_mk_r4_cls}
\title{Write scripts to make readyforwhatsnext S4 class}
\usage{
write_scripts_to_mk_r4_cls(
  name_stub_1L_chr,
  name_pfx_1L_chr = "ready4_",
  output_dir_1L_chr = "data-raw",
  outp_sub_dir_1L_chr = NULL,
  class_desc_1L_chr = "",
  parent_cls_nm_1L_chr = NULL,
  slots_chr,
  type_chr,
  meaningful_nms_ls = NULL,
  vals_ls = NULL,
  allowed_vals_ls = NULL,
  clss_to_inc_chr = NULL,
  prototype_lup,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  names_must_match_ls = NULL,
  slots_of_dif_lnts_chr = NULL,
  print_set_cls_1L_lgl = TRUE,
  print_helper = TRUE,
  print_gtrs_strs_1L_lgl = TRUE,
  print_validator_1L_lgl = TRUE,
  print_meaningful_nms_ls_1L_lgl = TRUE,
  class_in_cache_cdn_1L_chr = "stop"
)
}
\arguments{
\item{name_stub_1L_chr}{Name stub (a character vector of length one)}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one), Default: 'ready4_'}

\item{output_dir_1L_chr}{Output directory (a character vector of length one), Default: 'data-raw'}

\item{outp_sub_dir_1L_chr}{Output sub directory (a character vector of length one), Default: NULL}

\item{class_desc_1L_chr}{Class description (a character vector of length one), Default: ''}

\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one), Default: NULL}

\item{slots_chr}{Slots (a character vector)}

\item{type_chr}{Type (a character vector)}

\item{meaningful_nms_ls}{Meaningful names (a list), Default: NULL}

\item{vals_ls}{Values (a list), Default: NULL}

\item{allowed_vals_ls}{Allowed values (a list), Default: NULL}

\item{clss_to_inc_chr}{Classes to include (a character vector), Default: NULL}

\item{prototype_lup}{Prototype (a lookup table)}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector), Default: 'NA'}

\item{req_pkgs_chr}{Req packages (a character vector), Default: 'NA'}

\item{names_must_match_ls}{Names must match (a list), Default: NULL}

\item{slots_of_dif_lnts_chr}{Slots of different lengths (a character vector), Default: NULL}

\item{print_set_cls_1L_lgl}{Print set class (a logical vector of length one), Default: TRUE}

\item{print_helper}{PARAM_DESCRIPTION, Default: TRUE}

\item{print_gtrs_strs_1L_lgl}{Print getters setters (a logical vector of length one), Default: TRUE}

\item{print_validator_1L_lgl}{Print validator (a logical vector of length one), Default: TRUE}

\item{print_meaningful_nms_ls_1L_lgl}{Print meaningful names list (a logical vector of length one), Default: TRUE}

\item{class_in_cache_cdn_1L_chr}{Class in cache condition (a character vector of length one), Default: 'stop'}
}
\description{
write_scripts_to_mk_r4_cls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write scripts to make readyforwhatsnext s4 class.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
