% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scripts_to_mk_r4_clss}
\alias{write_scripts_to_mk_r4_clss}
\title{Write scripts to make readyforwhatsnext S4 classes}
\usage{
write_scripts_to_mk_r4_clss(
  name_stub,
  name_prefix = "ready4_",
  output_folder = "data-raw",
  output_sub_folder = NULL,
  class_desc = "",
  parent = NULL,
  class_slots,
  type,
  meaningful_names = NULL,
  values = NULL,
  allowed_values = NULL,
  include_classes = NULL,
  prototype_lup,
  ignore_ns_chr = NA_character_,
  required_pckg_chr_vec = NA_character_,
  names_include = NULL,
  not_same_length = NULL,
  print_set_class = TRUE,
  print_helper = TRUE,
  print_accessors = TRUE,
  print_validator = TRUE,
  print_meaningful_names = TRUE,
  class_in_cache_logic_chr = "stop"
)
}
\arguments{
\item{name_stub}{PARAM_DESCRIPTION}

\item{name_prefix}{PARAM_DESCRIPTION, Default: 'ready4_'}

\item{output_folder}{PARAM_DESCRIPTION, Default: 'data-raw'}

\item{output_sub_folder}{PARAM_DESCRIPTION, Default: NULL}

\item{class_desc}{PARAM_DESCRIPTION, Default: ''}

\item{parent}{PARAM_DESCRIPTION, Default: NULL}

\item{class_slots}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}

\item{meaningful_names}{PARAM_DESCRIPTION, Default: NULL}

\item{values}{PARAM_DESCRIPTION, Default: NULL}

\item{allowed_values}{PARAM_DESCRIPTION, Default: NULL}

\item{include_classes}{PARAM_DESCRIPTION, Default: NULL}

\item{prototype_lup}{Prototype (a lookup table)}

\item{ignore_ns_chr}{Ignore namespace (a character vector), Default: 'NA'}

\item{required_pckg_chr_vec}{PARAM_DESCRIPTION, Default: 'NA'}

\item{names_include}{PARAM_DESCRIPTION, Default: NULL}

\item{not_same_length}{PARAM_DESCRIPTION, Default: NULL}

\item{print_set_class}{PARAM_DESCRIPTION, Default: TRUE}

\item{print_helper}{PARAM_DESCRIPTION, Default: TRUE}

\item{print_accessors}{PARAM_DESCRIPTION, Default: TRUE}

\item{print_validator}{PARAM_DESCRIPTION, Default: TRUE}

\item{print_meaningful_names}{PARAM_DESCRIPTION, Default: TRUE}

\item{class_in_cache_logic_chr}{Class in cache logic (a character vector), Default: 'stop'}
}
\description{
write_scripts_to_mk_r4_clss() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write scripts to mk a readyforwhatsnext S4 clss.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
