% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_self_srvc_clss}
\alias{write_self_srvc_clss}
\title{Write self service classes}
\usage{
write_self_srvc_clss(pkg_setup_ls)
}
\arguments{
\item{pkg_setup_ls}{Package setup (a list)}
}
\value{
Package setup (a list)
}
\description{
write_self_srvc_clss() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write self service classes. The function returns Package setup (a list).
}
\keyword{internal}
