% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_slot_gtr_str_mthds}
\alias{write_slot_gtr_str_mthds}
\title{Write slot getter setter methods}
\usage{
write_slot_gtr_str_mthds(
  slot_name_chr,
  set_only,
  parent,
  class_name,
  print_accessors,
  output_folder,
  ignore_ns_chr,
  required_pckg_chr_vec
)
}
\arguments{
\item{slot_name_chr}{Slot name (a character vector)}

\item{set_only}{PARAM_DESCRIPTION}

\item{parent}{PARAM_DESCRIPTION}

\item{class_name}{PARAM_DESCRIPTION}

\item{print_accessors}{PARAM_DESCRIPTION}

\item{output_folder}{PARAM_DESCRIPTION}

\item{ignore_ns_chr}{Ignore namespace (a character vector)}

\item{required_pckg_chr_vec}{PARAM_DESCRIPTION}
}
\description{
write_slot_gtr_str_mthds() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write slot a getter setter mthds.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
