% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_std_method}
\alias{write_std_method}
\title{Write standard method}
\usage{
write_std_method(
  fn,
  fn_name_chr,
  class_chr,
  fn_desc_chr_vec,
  fn_title_chr,
  fn_out_type_chr,
  package_chr,
  output_dir_chr,
  signature_chr = NA_character_,
  append_lgl = T,
  first_lgl = T
)
}
\arguments{
\item{fn}{Function (a function)}

\item{fn_name_chr}{Function name (a character vector of length 1)}

\item{class_chr}{Class (a character vector of length 1)}

\item{fn_desc_chr_vec}{Function description (a character vector)}

\item{fn_title_chr}{Function title (a character vector of length 1)}

\item{fn_out_type_chr}{Function out type (a character vector of length 1)}

\item{package_chr}{Package (a character vector of length 1)}

\item{output_dir_chr}{Output directory (a character vector of length 1)}

\item{signature_chr}{Signature (a character vector of length 1), Default: 'NA'}

\item{append_lgl}{Append (a logical vector of length 1), Default: T}

\item{first_lgl}{First (a logical vector of length 1), Default: T}
}
\description{
write_std_method() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a standard method.NA
}
\keyword{internal}
