% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_std_mthd}
\alias{write_std_mthd}
\title{Write standard method}
\usage{
write_std_mthd(
  fn,
  fn_name_chr,
  class_chr,
  fn_desc_chr_vec,
  fn_title_chr,
  fn_out_type_chr,
  package_chr,
  output_dir_chr,
  signature_chr = NA_character_,
  append_lgl = T,
  first_lgl = T
)
}
\arguments{
\item{fn}{Function (a function)}

\item{fn_name_chr}{Function name (a character vector)}

\item{class_chr}{Class (a character vector)}

\item{fn_desc_chr_vec}{PARAM_DESCRIPTION}

\item{fn_title_chr}{Function title (a character vector)}

\item{fn_out_type_chr}{Function out type (a character vector)}

\item{package_chr}{Package (a character vector)}

\item{output_dir_chr}{Output directory (a character vector)}

\item{signature_chr}{Signature (a character vector), Default: 'NA'}

\item{append_lgl}{Append (a logical vector), Default: T}

\item{first_lgl}{First (a logical vector), Default: T}
}
\description{
write_std_mthd() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a standard mthd.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
