% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_delete_fls_with_ptrn}
\alias{write_to_delete_fls_with_ptrn}
\title{Write to delete files with pattern}
\usage{
write_to_delete_fls_with_ptrn(dir_1L_chr, pattern_1L_chr)
}
\arguments{
\item{dir_1L_chr}{Directory (a character vector of length one)}

\item{pattern_1L_chr}{Pattern (a character vector of length one)}
}
\description{
write_to_delete_fls_with_ptrn() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to delete files with pattern. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
