% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_mk_r4_cls}
\alias{write_to_mk_r4_cls}
\title{Write to make readyforwhatsnext S4 class}
\usage{
write_to_mk_r4_cls(
  class_name,
  class_slots,
  type,
  proto_ls,
  parent,
  print_set_class,
  class_desc,
  output_file_class,
  include_classes,
  prototype_lup,
  helper_lgl = F,
  parent_ns_ls
)
}
\arguments{
\item{class_name}{PARAM_DESCRIPTION}

\item{class_slots}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}

\item{proto_ls}{Proto (a list)}

\item{parent}{PARAM_DESCRIPTION}

\item{print_set_class}{PARAM_DESCRIPTION}

\item{class_desc}{PARAM_DESCRIPTION}

\item{output_file_class}{PARAM_DESCRIPTION}

\item{include_classes}{PARAM_DESCRIPTION}

\item{prototype_lup}{Prototype (a lookup table)}

\item{helper_lgl}{Helper (a logical vector), Default: F}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
write_to_mk_r4_cls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to mk a readyforwhatsnext S4 cls.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
