% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_add_class.R
\name{add_class.ready4_class_pt_lup}
\alias{add_class.ready4_class_pt_lup}
\title{Add class method applied to Class Prototype Lookup Table readyforwhatsnext S3 class.}
\usage{
\method{add_class}{ready4_class_pt_lup}(
  x,
  row_idx_1L_int,
  make_tb,
  dev_pkg_ns_1L_chr,
  name_pfx_1L_chr,
  output_dir_1L_chr,
  file_exists_cdn_1L_chr,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup = NULL
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{row_idx_1L_int}{Row idx (an integer vector of length one)}

\item{make_tb}{Make (a tibble)}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one)}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one)}

\item{file_exists_cdn_1L_chr}{File exists condition (a character vector of length one)}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector), Default: 'NA'}

\item{req_pkgs_chr}{Req packages (a character vector), Default: 'NA'}

\item{class_in_cache_cdn_1L_chr}{Class in cache condition (a character vector of length one), Default: 'stop'}

\item{abbreviations_lup}{Abbreviations (a lookup table), Default: NULL}
}
\value{
Inst of ready4 class prototype (a lookup table)
}
\description{
add_class.ready4_class_pt_lup() is an Add Class method that adds information about a class. This method is implemented for the Class Prototype Lookup Table readyforwhatsnext S3 class.The function returns inst of ready4 class prototype (a lookup table).
}
\keyword{internal}
