% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author.ready4class_constructor}
\alias{author.ready4class_constructor}
\alias{author,ready4class_constructor-method}
\alias{author.ready4class_manifest}
\alias{author,ready4class_manifest-method}
\alias{author.ready4class_pt_lup}
\alias{author,ready4class_pt_lup-method}
\title{Author method applied to ready4 S3 class Constructor Table..}
\usage{
author.ready4class_constructor(
  x,
  dev_pkg_ns_1L_chr = ready4fun::get_dev_pkg_nm(),
  name_pfx_1L_chr = paste0(ready4fun::get_dev_pkg_nm(), "_"),
  output_dir_1L_chr = "R",
  delete_cdn_ptrn_chr = NA_character_,
  file_exists_cdn_1L_chr = "overwrite",
  init_class_pt_lup = NULL,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup,
  object_type_lup
)

\S4method{author}{ready4class_constructor}(
  x,
  dev_pkg_ns_1L_chr = ready4fun::get_dev_pkg_nm(),
  name_pfx_1L_chr = paste0(ready4fun::get_dev_pkg_nm(), "_"),
  output_dir_1L_chr = "R",
  delete_cdn_ptrn_chr = NA_character_,
  file_exists_cdn_1L_chr = "overwrite",
  init_class_pt_lup = NULL,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup,
  object_type_lup
)

author.ready4class_manifest(
  x,
  init_class_pt_lup = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  list_generics_1L_lgl = F,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL
)

\S4method{author}{ready4class_manifest}(
  x,
  init_class_pt_lup = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  list_generics_1L_lgl = F,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  self_serve_1L_lgl = F,
  self_serve_fn_ls = NULL
)

author.ready4class_pt_lup(
  x,
  row_idx_1L_int,
  make_tb,
  dev_pkg_ns_1L_chr,
  name_pfx_1L_chr,
  output_dir_1L_chr,
  file_exists_cdn_1L_chr,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup,
  object_type_lup,
  consent_1L_chr = NULL
)

\S4method{author}{ready4class_pt_lup}(
  x,
  row_idx_1L_int,
  make_tb,
  dev_pkg_ns_1L_chr,
  name_pfx_1L_chr,
  output_dir_1L_chr,
  file_exists_cdn_1L_chr,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_,
  class_in_cache_cdn_1L_chr = "stop",
  abbreviations_lup,
  object_type_lup,
  consent_1L_chr = NULL
)
}
\arguments{
\item{x}{An instance of ready4 S3 class Prototype Lookup Table of class metadata.}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one)}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one)}

\item{delete_cdn_ptrn_chr}{Delete condition pattern (a character vector), Default: 'NA'}

\item{file_exists_cdn_1L_chr}{File exists condition (a character vector of length one)}

\item{init_class_pt_lup}{Initial class prototype (a lookup table), Default: NULL}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector), Default: 'NA'}

\item{req_pkgs_chr}{Require packages (a character vector), Default: 'NA'}

\item{class_in_cache_cdn_1L_chr}{Class in cache condition (a character vector of length one), Default: 'stop'}

\item{abbreviations_lup}{Abbreviations (a lookup table)}

\item{object_type_lup}{Object type (a lookup table)}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{list_generics_1L_lgl}{List generics (a logical vector of length one), Default: F}

\item{self_serve_1L_lgl}{Self serve (a logical vector of length one), Default: F}

\item{self_serve_fn_ls}{Self serve (a list of functions), Default: NULL}

\item{row_idx_1L_int}{Row index (an integer vector of length one)}

\item{make_tb}{Make (a tibble)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: NULL}
}
\value{
Instance (ready4 S3 class Prototype Lookup Table of class metadata.)

X (ready4 S3 class for encapsulating the metadata required for package set-up.)

Instance (ready4 S3 class Prototype Lookup Table of class metadata.)
}
\description{
author.ready4class_constructor() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 S3 class Constructor Table. The function returns Instance (ready4 S3 class Prototype Lookup Table of class metadata.).

author.ready4class_manifest() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 S3 class Manifest. The function is called for its side effects and does not return a value.

author.ready4class_pt_lup() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 S3 class Prototype Lookup Table of class metadata. The function returns Instance (ready4 S3 class Prototype Lookup Table of class metadata.).
}
