% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_class_ns}
\alias{get_class_ns}
\title{Get class namespace}
\usage{
get_class_ns(prototype_lup, class_nm_1L_chr)
}
\arguments{
\item{prototype_lup}{Prototype (a lookup table)}

\item{class_nm_1L_chr}{Class name (a character vector of length one)}
}
\value{
Class namespace (a character vector of length one)
}
\description{
get_class_ns() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get class namespace. Function argument prototype_lup specifies the where to look for the required object.The function returns a class namespace (a character vector of length one).
}
