% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_nms_of_clss_to_inc}
\alias{get_nms_of_clss_to_inc}
\title{Get names of classes to include}
\usage{
get_nms_of_clss_to_inc(
  parent_cls_nm_1L_chr,
  parent_ns_ls,
  base_set_of_clss_to_inc_chr = NULL
)
}
\arguments{
\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{parent_ns_ls}{Parent namespace (a list)}

\item{base_set_of_clss_to_inc_chr}{Base set of classes to include (a character vector), Default: NULL}
}
\value{
Names of classes to include (a character vector)
}
\description{
get_nms_of_clss_to_inc() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get names of classes to include. The function returns Names of classes to include (a character vector).
}
\keyword{internal}
