% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_nms_of_clss_to_inc}
\alias{get_nms_of_clss_to_inc}
\title{Get names of classes to include}
\usage{
get_nms_of_clss_to_inc(
  parent_cls_nm_1L_chr,
  parent_ns_ls,
  base_set_of_clss_to_inc_chr = NULL
)
}
\arguments{
\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{parent_ns_ls}{Parent namespace (a list)}

\item{base_set_of_clss_to_inc_chr}{Base set of classes to include (a character vector), Default: NULL}
}
\value{
Names of clss to inc (a character vector)
}
\description{
get_nms_of_clss_to_inc() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get names of classes to include. Function argument parent_cls_nm_1L_chr specifies the where to look for the required object.The function returns names of clss to inc (a character vector).
}
