% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_nms_of_curr_gnrcs}
\alias{get_nms_of_curr_gnrcs}
\title{Get names of current generics}
\usage{
get_nms_of_curr_gnrcs(required_pckg_chr_vec, generic_chr)
}
\arguments{
\item{required_pckg_chr_vec}{PARAM_DESCRIPTION}

\item{generic_chr}{Generic (a character vector)}
}
\description{
get_nms_of_curr_gnrcs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get names of curr gnrcs. Function argument required_pckg_chr_vec specifies the where to look for the required object.The function is called for its side effects and does not return a value.
}
