% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parent_cls_ns}
\alias{get_parent_cls_ns}
\title{Get parent class namespace}
\usage{
get_parent_cls_ns(prototype_lup, parent_chr, dev_pckg_ns_chr)
}
\arguments{
\item{prototype_lup}{Prototype (a lookup table)}

\item{parent_chr}{Parent (a character vector)}

\item{dev_pckg_ns_chr}{Development pckg namespace (a character vector)}
}
\description{
get_parent_cls_ns() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get parent cls a namespace. Function argument prototype_lup specifies the where to look for the required object.The function is called for its side effects and does not return a value.
}
\keyword{internal}
