% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parent_cls_pt_fn}
\alias{get_parent_cls_pt_fn}
\title{Get parent class prototype}
\usage{
get_parent_cls_pt_fn(parent_cls_nm_1L_chr, prototype_lup)
}
\arguments{
\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{prototype_lup}{Prototype (a lookup table)}
}
\value{
Parent cls prototype (a function)
}
\description{
get_parent_cls_pt_fn() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get parent class prototype function. Function argument parent_cls_nm_1L_chr specifies the where to look for the required object. The function returns parent class prototype (a function).
}
