% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parent_cls_pts}
\alias{get_parent_cls_pts}
\title{Get parent class prototypes}
\usage{
get_parent_cls_pts(parent_cls_nm_1L_chr, parent_ns_ls, slot_names_chr)
}
\arguments{
\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{parent_ns_ls}{Parent namespace (a list)}

\item{slot_names_chr}{Slot names (a character vector)}
}
\value{
Parent cls prototypes (a character vector)
}
\description{
get_parent_cls_pts() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get parent cls prototypes. Function argument parent_cls_nm_1L_chr specifies the where to look for the required object.The function returns parent cls prototypes (a character vector).
}
\keyword{internal}
