% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_parent_cls_slot_nms}
\alias{get_parent_cls_slot_nms}
\title{Get parent class slot names}
\usage{
get_parent_cls_slot_nms(parent_cls_nm_1L_chr, parent_ns_ls)
}
\arguments{
\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\value{
Parent class slot names (a character vector)
}
\description{
get_parent_cls_slot_nms() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get parent class slot names. Function argument parent_cls_nm_1L_chr specifies the where to look for the required object. The function returns Parent class slot names (a character vector).
}
\keyword{internal}
