% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_gen_ref_to_cls}
\alias{make_alg_to_gen_ref_to_cls}
\title{Make algorithm to generate reference to class}
\usage{
make_alg_to_gen_ref_to_cls(class_nm_1L_chr, pkg_nm_1L_chr = ".GlobalEnv")
}
\arguments{
\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: '.GlobalEnv'}
}
\value{
Algorithm to generate reference to class (a character vector of length one)
}
\description{
make_alg_to_gen_ref_to_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make algorithm to generate reference to class. The function returns Algorithm to generate reference to class (a character vector of length one).
}
