% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_get_pt_val}
\alias{make_alg_to_get_pt_val}
\title{Make algorithm to get prototype value}
\usage{
make_alg_to_get_pt_val(
  type_namespace = "",
  function_to_call = "",
  default_value = "",
  namespace_contexts = c("base")
)
}
\arguments{
\item{type_namespace}{PARAM_DESCRIPTION, Default: ''}

\item{function_to_call}{PARAM_DESCRIPTION, Default: ''}

\item{default_value}{PARAM_DESCRIPTION, Default: ''}

\item{namespace_contexts}{PARAM_DESCRIPTION, Default: c("base")}
}
\description{
make_alg_to_get_pt_val() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make alg to get a prototype val.The function is called for its side effects and does not return a value.
}
