% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_set_gnrc}
\alias{make_alg_to_set_gnrc}
\title{Make algorithm to set generic}
\usage{
make_alg_to_set_gnrc(
  name_1L_chr,
  args_chr = c("x"),
  signature_1L_chr = NA_character_,
  where_1L_chr = NA_character_
)
}
\arguments{
\item{name_1L_chr}{Name (a character vector of length one)}

\item{args_chr}{Arguments (a character vector), Default: c("x")}

\item{signature_1L_chr}{Signature (a character vector of length one), Default: 'NA'}

\item{where_1L_chr}{Where (a character vector of length one), Default: 'NA'}
}
\value{
Algorithm to set generic (a character vector of length one)
}
\description{
make_alg_to_set_gnrc() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make algorithm to set generic. The function returns Algorithm to set generic (a character vector of length one).
}
