% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_set_mthd}
\alias{make_alg_to_set_mthd}
\title{Make algorithm to set method}
\usage{
make_alg_to_set_mthd(
  name_1L_chr,
  class_nm_1L_chr,
  fn = NULL,
  fn_nm_1L_chr = NA_character_,
  pkg_nm_1L_chr = NA_character_,
  where_1L_chr = NA_character_
)
}
\arguments{
\item{name_1L_chr}{Name (a character vector of length one)}

\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{fn}{Function (a function), Default: NULL}

\item{fn_nm_1L_chr}{Function name (a character vector of length one), Default: 'NA'}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: 'NA'}

\item{where_1L_chr}{Where (a character vector of length one), Default: 'NA'}
}
\value{
Algorithm to set method (a character vector of length one)
}
\description{
make_alg_to_set_mthd() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make algorithm to set method. The function returns Algorithm to set method (a character vector of length one).
}
\keyword{internal}
