% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_set_old_clss}
\alias{make_alg_to_set_old_clss}
\title{Make algorithm to set old classes}
\usage{
make_alg_to_set_old_clss(type, prototype_lup)
}
\arguments{
\item{type}{PARAM_DESCRIPTION}

\item{prototype_lup}{Prototype (a lookup table)}
}
\description{
make_alg_to_set_old_clss() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make alg to set old clss.The function is called for its side effects and does not return a value.
}
\keyword{internal}
