% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_set_validity_of_r4_cls}
\alias{make_alg_to_set_validity_of_r4_cls}
\title{Make algorithm to set validity of readyforwhatsnext S4 class}
\usage{
make_alg_to_set_validity_of_r4_cls(
  class_nm_1L_chr,
  parent_cls_nm_1L_chr,
  slots_of_dif_lnts_chr = NULL,
  allowed_vals_ls = NULL,
  names_must_match_ls = NULL,
  print_validator_1L_lgl = FALSE
)
}
\arguments{
\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{slots_of_dif_lnts_chr}{Slots of different lengths (a character vector), Default: NULL}

\item{allowed_vals_ls}{Allowed values (a list), Default: NULL}

\item{names_must_match_ls}{Names must match (a list), Default: NULL}

\item{print_validator_1L_lgl}{Print validator (a logical vector of length one), Default: FALSE}
}
\value{
Alg to set validity of readyforwhatsnext S4 cls (a character vector of length one)
}
\description{
make_alg_to_set_validity_of_r4_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make algorithm to set validity of readyforwhatsnext s4 class. The function returns an algorithm to set validity of readyforwhatsnext s4 class (a character vector of length one).
}
