% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_alg_to_write_gtr_str_mthds}
\alias{make_alg_to_write_gtr_str_mthds}
\title{Make algorithm to write getter setter methods}
\usage{
make_alg_to_write_gtr_str_mthds(
  class_nm_1L_chr,
  parent_cls_nm_1L_chr,
  print_gtrs_strs_1L_lgl,
  output_dir_1L_chr,
  nss_to_ignore_chr,
  req_pkgs_chr,
  parent_ns_ls
)
}
\arguments{
\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{print_gtrs_strs_1L_lgl}{Print getters setters (a logical vector of length one)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one)}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector)}

\item{req_pkgs_chr}{Req packages (a character vector)}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
make_alg_to_write_gtr_str_mthds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make algorithm to write getter setter methods. The function is called for its side effects and does not return a value.
}
