% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{make_and_update}
\alias{make_and_update}
\title{Make and update}
\usage{
make_and_update(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\description{
make_and_update() is a Make and Update generic that applies a Make method and then updates the output of that method. The function is called for its side effects and does not return a value.
}
\keyword{internal}
