% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_and_update.R
\name{make_and_update.ready4_class_make_tb}
\alias{make_and_update.ready4_class_make_tb}
\title{Make and update method applied to Class Make Table readyforwhatsnext S3 class.}
\usage{
\method{make_and_update}{ready4_class_make_tb}(
  x,
  dev_pckg_namespace = ready4fun::get_dev_pkg_nm_1L_chr(),
  name_prefix = paste0(ready4fun::get_dev_pkg_nm_1L_chr(), "_"),
  output_dir = "R",
  delete_files_pattern_chr_vec = NA_character_,
  file_exists_logic = "overwrite",
  init_class_pt_lup = NULL,
  ignore_ns_chr = NA_character_,
  required_pckg_chr_vec = NA_character_,
  class_in_cache_logic_chr = "stop"
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{dev_pckg_namespace}{PARAM_DESCRIPTION, Default: ready4fun::get_dev_pkg_nm_1L_chr()}

\item{name_prefix}{PARAM_DESCRIPTION, Default: paste0(ready4fun::get_dev_pkg_nm_1L_chr(), "_")}

\item{output_dir}{PARAM_DESCRIPTION, Default: 'R'}

\item{delete_files_pattern_chr_vec}{Delete files pattern (a character vector), Default: 'NA'}

\item{file_exists_logic}{PARAM_DESCRIPTION, Default: 'overwrite'}

\item{init_class_pt_lup}{Init class prototype (a lookup table), Default: NULL}

\item{ignore_ns_chr}{Ignore namespace (a character vector of length 1), Default: 'NA'}

\item{required_pckg_chr_vec}{Required pckg (a character vector), Default: 'NA'}

\item{class_in_cache_logic_chr}{Class in cache logic (a character vector of length 1), Default: 'stop'}
}
\description{
make_and_update.ready4_class_make_tb() is a Make and Update method that applies a Make method and then updates the output of that method. This method is implemented for the Class Make Table readyforwhatsnext S3 class.NA
}
\keyword{internal}
