% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_child_cls_fn_body}
\alias{make_child_cls_fn_body}
\title{Make child class function body}
\usage{
make_child_cls_fn_body(
  child_ext_fn_chr,
  parent_chr,
  prototype_lup,
  prepend_lgl = T
)
}
\arguments{
\item{child_ext_fn_chr}{Child ext function (a character vector)}

\item{parent_chr}{Parent (a character vector)}

\item{prototype_lup}{Prototype (a lookup table)}

\item{prepend_lgl}{Prepend (a logical vector), Default: T}
}
\description{
make_child_cls_fn_body() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make child cls a function body.The function is called for its side effects and does not return a value.
}
\keyword{internal}
