% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_class_pt_tb_for_r3_and_r4_clss}
\alias{make_class_pt_tb_for_r3_and_r4_clss}
\title{Make class prototype tibble for readyforwhatsnext S3 and readyforwhatsnext S4 classes}
\usage{
make_class_pt_tb_for_r3_and_r4_clss(class_mk_ls)
}
\arguments{
\item{class_mk_ls}{Class make (a list)}
}
\value{
Class prototype tibble for readyforwhatsnext S3 and readyforwhatsnext S4 clss (a tibble)
}
\description{
make_class_pt_tb_for_r3_and_r4_clss() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make class prototype tibble for a readyforwhatsnext S3 and a readyforwhatsnext S4 clss.The function returns a class prototype tibble for readyforwhatsnext s3 and readyforwhatsnext s4 clss (a tibble).
}
\keyword{internal}
