% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_classes.R
\name{make_classes.ready4_class_make_tb}
\alias{make_classes.ready4_class_make_tb}
\title{Make classes method applied to Class Make Table readyforwhatsnext S3 class.}
\usage{
\method{make_classes}{ready4_class_make_tb}(
  x,
  name_prefix,
  output_dir,
  file_exists_logic = NULL,
  prototype_lup = NULL,
  ignore_ns_chr,
  required_pckg_chr_vec = NA_character_,
  class_in_cache_logic_chr = "stop"
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{name_prefix}{PARAM_DESCRIPTION}

\item{output_dir}{PARAM_DESCRIPTION}

\item{file_exists_logic}{PARAM_DESCRIPTION, Default: NULL}

\item{prototype_lup}{Prototype (a lookup table), Default: NULL}

\item{ignore_ns_chr}{Ignore namespace (a character vector)}

\item{required_pckg_chr_vec}{PARAM_DESCRIPTION, Default: 'NA'}

\item{class_in_cache_logic_chr}{Class in cache logic (a character vector), Default: 'stop'}
}
\description{
make_classes.ready4_class_make_tb() is a Make Classes method that writes new classes. This method is implemented for the Class Make Table readyforwhatsnext S3 class.NA
}
