% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_dmt_inc_tag}
\alias{make_dmt_inc_tag}
\title{Make documentation inc tag}
\usage{
make_dmt_inc_tag(class_names_chr, s3_1L_lgl = T)
}
\arguments{
\item{class_names_chr}{Class names (a character vector)}

\item{s3_1L_lgl}{S3 (a logical vector of length one), Default: T}
}
\value{
Documentation inc tag (a character vector of length one)
}
\description{
make_dmt_inc_tag() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make documentation inc tag.The function returns documentation inc tag (a character vector of length one).
}
\keyword{internal}
