% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gnrc_fn}
\alias{make_gnrc_fn}
\title{Make generic}
\usage{
make_gnrc_fn(name_1L_chr, args_chr)
}
\arguments{
\item{name_1L_chr}{Name (a character vector of length one)}

\item{args_chr}{Arguments (a character vector)}
}
\value{
Gnrc function (a character vector of length one)
}
\description{
make_gnrc_fn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make gnrc.The function returns gnrc function (a character vector of length one).
}
