% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gnrc_mthd_pair_ls}
\alias{make_gnrc_mthd_pair_ls}
\title{Make generic method pair}
\usage{
make_gnrc_mthd_pair_ls(
  name_1L_chr,
  args_chr = c("x"),
  signature_1L_chr = NA_character_,
  pkg_nm_1L_chr = NA_character_,
  where_1L_chr = NA_character_,
  class_nm_1L_chr,
  fn
)
}
\arguments{
\item{name_1L_chr}{Name (a character vector of length one)}

\item{args_chr}{Arguments (a character vector), Default: c("x")}

\item{signature_1L_chr}{Signature (a character vector of length one), Default: 'NA'}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: 'NA'}

\item{where_1L_chr}{Where (a character vector of length one), Default: 'NA'}

\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{fn}{Function (a function)}
}
\value{
Gnrc mthd pair (a list)
}
\description{
make_gnrc_mthd_pair_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make generic method pair list.The function returns gnrc mthd pair (a list).
}
