% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ls_of_pkgs_to_imp}
\alias{make_ls_of_pkgs_to_imp}
\title{Make list of packages to import}
\usage{
make_ls_of_pkgs_to_imp(curr_gnrcs_ls, fn_name_1L_chr, nss_to_ignore_chr)
}
\arguments{
\item{curr_gnrcs_ls}{Current generics (a list)}

\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector)}
}
\value{
Packages to import (a list)
}
\description{
make_ls_of_pkgs_to_imp() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make list of packages to an import.The function returns packages to import (a list).
}
\keyword{internal}
