% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{make_lup}
\alias{make_lup}
\title{Make lookup table}
\usage{
make_lup(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
make_lup() is a Make Lookup Table generic that applies a Make method and then updates the output of that method.NA
}
\keyword{internal}
