% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_lup.R
\name{make_lup.ready4_class_make_tb}
\alias{make_lup.ready4_class_make_tb}
\title{Make lookup table method applied to Class Make Table readyforwhatsnext S3 class.}
\usage{
\method{make_lup}{ready4_class_make_tb}(x, dev_pkg_ns_1L_chr, prefix_1L_chr)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}
}
\value{
Inst of ready4 class prototype (a lookup table)
}
\description{
make_lup.ready4_class_make_tb() is a Make Lookup Table method that makes a lookup table. This method is implemented for the Class Make Table readyforwhatsnext S3 class.The function returns inst of ready4 class prototype (a lookup table).
}
\keyword{internal}
