% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{make_new_ready4_class_pt_lup}
\alias{make_new_ready4_class_pt_lup}
\title{Make new CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class}
\usage{
make_new_ready4_class_pt_lup(x)
}
\arguments{
\item{x}{A prototype for the CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class}
}
\value{
An unvalidated instance of the CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class
}
\description{
Create a new unvalidated instance of the CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class
}
\details{
PROTOTYPE LOOKUP Readyforwhatsnext S3 class of metadata of prototype classes.
}
