% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_constructor_tbl.R
\name{make_new_ready4_constructor_tbl}
\alias{make_new_ready4_constructor_tbl}
\title{Make new CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class}
\usage{
make_new_ready4_constructor_tbl(x)
}
\arguments{
\item{x}{A prototype for the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class}
}
\value{
An unvalidated instance of the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\description{
Create a new unvalidated instance of the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\details{
CLASS CONSTRUCTOR TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
