% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor.R
\name{make_new_ready4class_constructor}
\alias{make_new_ready4class_constructor}
\title{make new ready4class constructor Class constructor table}
\usage{
make_new_ready4class_constructor(x)
}
\arguments{
\item{x}{A prototype for the Class constructor table}
}
\value{
An unvalidated instance of the Class constructor table
}
\description{
Create a new unvalidated instance of the Class constructor table
}
\details{
Class constructor table
}
\keyword{internal}
