% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_manifest.R
\name{make_new_ready4class_manifest}
\alias{make_new_ready4class_manifest}
\title{Make new ready4class package manifest ready4 S3 class Manifest.}
\usage{
make_new_ready4class_manifest(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class Manifest.}
}
\value{
An unvalidated instance of the ready4 S3 class Manifest.
}
\description{
Create a new unvalidated instance of the ready4 S3 class Manifest.
}
\details{
ready4 S3 class Manifest.
}
