% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_pt_lup.R
\name{make_new_ready4class_pt_lup}
\alias{make_new_ready4class_pt_lup}
\title{make new ready4class package prototype lookup table ready4 S3 class Prototype Lookup Table of class metadata.}
\usage{
make_new_ready4class_pt_lup(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class Prototype Lookup Table of class metadata.}
}
\value{
An unvalidated instance of the ready4 S3 class Prototype Lookup Table of class metadata.
}
\description{
Create a new unvalidated instance of the ready4 S3 class Prototype Lookup Table of class metadata.
}
\details{
ready4 S3 class Prototype Lookup Table.
}
\keyword{internal}
