% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_one_row_class_pt_tb}
\alias{make_one_row_class_pt_tb}
\title{Make one row class prototype tibble}
\usage{
make_one_row_class_pt_tb(class_type_mk_ls, make_s3_1L_lgl = T)
}
\arguments{
\item{class_type_mk_ls}{Class type make (a list)}

\item{make_s3_1L_lgl}{Make S3 (a logical vector of length one), Default: T}
}
\value{
One row class prototype (a tibble)
}
\description{
make_one_row_class_pt_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make one row class prototype tibble. The function returns One row class prototype (a tibble).
}
