% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_one_row_pt_tb_for_new_r3_cls}
\alias{make_one_row_pt_tb_for_new_r3_cls}
\title{Make one row prototype tibble for new readyforwhatsnext S3 class}
\usage{
make_one_row_pt_tb_for_new_r3_cls(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
make_one_row_pt_tb_for_new_r3_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make one row a prototype tibble for new a readyforwhatsnext S3 cls.The function is called for its side effects and does not return a value.
}
\keyword{internal}
