% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{make_prototype_ready4_class_pt_lup}
\alias{make_prototype_ready4_class_pt_lup}
\title{NA}
\usage{
make_prototype_ready4_class_pt_lup()
}
\value{
A prototype for ready4_class_pt_lup class
}
\description{
Create a new prototype for S3 class: ready4_class_pt_lup
}
\details{
PROTOTYPE LOOKUP Readyforwhatsnext S3 class of metadata of prototype classes.
}
