% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{make_prototype_ready4_class_pt_lup}
\alias{make_prototype_ready4_class_pt_lup}
\title{Make prototype CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class}
\usage{
make_prototype_ready4_class_pt_lup()
}
\value{
A prototype for CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class
}
\description{
Create a new prototype for the CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class
}
\details{
CLASS PROTOTYPE LOOKUP TABLE Readyforwhatsnext S3 class of metadata of prototype classes.
}
