% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_constructor_tbl.R
\name{make_prototype_ready4_constructor_tbl}
\alias{make_prototype_ready4_constructor_tbl}
\title{Make prototype CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class}
\usage{
make_prototype_ready4_constructor_tbl()
}
\value{
A prototype for CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\description{
Create a new prototype for the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\details{
CLASS CONSTRUCTOR TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
