% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pt_ls}
\alias{make_pt_ls}
\title{Make prototype}
\usage{
make_pt_ls(
  slots_chr,
  type_chr = NULL,
  vals_ls = NULL,
  make_val_1L_lgl = TRUE,
  prototype_lup
)
}
\arguments{
\item{slots_chr}{Slots (a character vector)}

\item{type_chr}{Type (a character vector), Default: NULL}

\item{vals_ls}{Values (a list), Default: NULL}

\item{make_val_1L_lgl}{Make value (a logical vector of length one), Default: TRUE}

\item{prototype_lup}{Prototype (a lookup table)}
}
\value{
Prototype (a list)
}
\description{
make_pt_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make prototype list.The function returns a prototype (a list).
}
