% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pt_ls_for_new_r3_cls}
\alias{make_pt_ls_for_new_r3_cls}
\title{Make prototype for new ready4 S3 class}
\usage{
make_pt_ls_for_new_r3_cls(
  class_name_1L_chr,
  type_1L_chr,
  pt_ns_1L_chr,
  pt_chkr_pfx_1L_chr,
  vals_ls,
  ordered_1L_lgl,
  parent_cls_nm_1L_chr,
  prototype_lup,
  min_max_vals_dbl,
  start_end_vals_dbl,
  dev_pkg_ns_1L_chr = ready4fun::get_dev_pkg_nm(),
  nss_to_ignore_chr,
  asserts_ls = NULL
)
}
\arguments{
\item{class_name_1L_chr}{Class name (a character vector of length one)}

\item{type_1L_chr}{Type (a character vector of length one)}

\item{pt_ns_1L_chr}{Prototype namespace (a character vector of length one)}

\item{pt_chkr_pfx_1L_chr}{Prototype checker prefix (a character vector of length one)}

\item{vals_ls}{Values (a list)}

\item{ordered_1L_lgl}{Ordered (a logical vector of length one)}

\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{prototype_lup}{Prototype (a lookup table)}

\item{min_max_vals_dbl}{Min max values (a double vector)}

\item{start_end_vals_dbl}{Start end values (a double vector)}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one), Default: ready4fun::get_dev_pkg_nm()}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector)}

\item{asserts_ls}{Asserts (a list), Default: NULL}
}
\value{
Prototype list for new ready4 S3 class (a list)
}
\description{
make_pt_ls_for_new_r3_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make prototype list for new ready4 s3 class. The function returns Prototype list for new ready4 S3 class (a list).
}
\keyword{internal}
