% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_manifest.R
\name{make_pt_ready4class_manifest}
\alias{make_pt_ready4class_manifest}
\title{Make prototype ready4class package manifest ready4 S3 class Manifest.}
\usage{
make_pt_ready4class_manifest(
  manifest_r3 = list(),
  constructor_r3 = ready4class_constructor()
)
}
\arguments{
\item{manifest_r3}{Manifest (a ready4 S3), Default: list()}

\item{constructor_r3}{Constructor (a ready4 S3), Default: ready4class_constructor()}
}
\value{
A prototype for ready4 S3 class Manifest.
}
\description{
Create a new prototype for the ready4 S3 class Manifest.
}
\details{
ready4 S3 class Manifest.
}
