% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pt_tb_for_new_r4_cls}
\alias{make_pt_tb_for_new_r4_cls}
\title{Make prototype tibble for new readyforwhatsnext S4 class}
\usage{
make_pt_tb_for_new_r4_cls(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
make_pt_tb_for_new_r4_cls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to a make prototype tibble for new a readyforwhatsnext S4 cls.The function is called for its side effects and does not return a value.
}
\keyword{internal}
