% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_show_mthd_fn}
\alias{make_show_mthd_fn}
\title{Make show method function}
\usage{
make_show_mthd_fn(class_nm_1L_chr, meaningful_nms_ls)
}
\arguments{
\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{meaningful_nms_ls}{Meaningful names (a list)}
}
\value{
Show method function (a character vector of length one)
}
\description{
make_show_mthd_fn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make show method function. The function returns Show method function (a character vector of length one).
}
