% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture.ready4class_constructor}
\alias{manufacture.ready4class_constructor}
\alias{manufacture,ready4class_constructor-method}
\title{manufacture - a method that manufactures a novel r object using data contained in an instance of a class}
\usage{
manufacture.ready4class_constructor(x, dev_pkg_ns_1L_chr, prefix_1L_chr)

\S4method{manufacture}{ready4class_constructor}(x, dev_pkg_ns_1L_chr, prefix_1L_chr)
}
\arguments{
\item{x}{An instance of ready4 S3 class Constructor Table.}

\item{dev_pkg_ns_1L_chr}{Development package namespace (a character vector of length one)}

\item{prefix_1L_chr}{Prefix (a character vector of length one)}
}
\value{
Instance (ready4 S3 class Prototype Lookup Table of class metadata.)
}
\description{
manufacture.ready4class_constructor() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 S3 class Constructor Table. The function returns Instance (ready4 S3 class Prototype Lookup Table of class metadata.).
}
