% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_make_tb.R
\name{ready4_class_make_tb}
\alias{ready4_class_make_tb}
\title{NA}
\usage{
ready4_class_make_tb(x = make_prototype_ready4_class_make_tb())
}
\arguments{
\item{x}{A prototype for the ready4_class_make_tb class, Default: make_prototype_ready4_class_make_tb()}
}
\value{
A validated instance of the ready4_class_make_tb class
}
\description{
Create a new valid instance of the S3 class: ready4_class_make_tb
}
\details{
MAKE CLASS TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
