% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{ready4_class_pt_lup}
\alias{ready4_class_pt_lup}
\title{readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE}
\usage{
ready4_class_pt_lup(x = make_prototype_ready4_class_pt_lup())
}
\arguments{
\item{x}{A prototype for the readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE, Default: make_prototype_ready4_class_pt_lup()}
}
\value{
A validated instance of the readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE
}
\description{
Create a new valid instance of the readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE
}
\details{
Readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE class of metadata of prototype classes.
}
