% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{ready4_class_pt_lup}
\alias{ready4_class_pt_lup}
\title{CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class}
\usage{
ready4_class_pt_lup(x = make_prototype_ready4_class_pt_lup())
}
\arguments{
\item{x}{A prototype for the CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class, Default: make_prototype_ready4_class_pt_lup()}
}
\value{
A validated instance of the CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class
}
\description{
Create a new valid instance of the CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class
}
\details{
PROTOTYPE LOOKUP Readyforwhatsnext S3 class of metadata of prototype classes.
}
