% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_constructor_tbl.R
\name{ready4_constructor_tbl}
\alias{ready4_constructor_tbl}
\title{CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class}
\usage{
ready4_constructor_tbl(x = make_prototype_ready4_constructor_tbl())
}
\arguments{
\item{x}{A prototype for the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class, Default: make_prototype_ready4_constructor_tbl()}
}
\value{
A validated instance of the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\description{
Create a new valid instance of the CLASS CONSTRUCTOR TABLE readyforwhatsnext S3 class
}
\details{
CLASS CONSTRUCTOR TABLE Readyforwhatsnext S3 class of metadata required to make new classes.
}
