% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4class.R
\docType{package}
\name{ready4class-package}
\alias{ready4class}
\alias{ready4class-package}
\title{ready4class: Readyforwhatsnext Tools For Creating And Extending Classes.}
\description{
Provides classes and methods that are designed to standardise and streamline the process for extending the readyforwhatsnext data synthesis and simulation framework with new classes.
}
\details{
To learn more about ready4class, start with the vignettes:
`browseVignettes(package = "ready4class")`
}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
