% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor.R
\name{ready4class_constructor}
\alias{ready4class_constructor}
\alias{make_pt_ready4class_constructor}
\alias{is_ready4class_constructor}
\title{Class constructor table}
\usage{
ready4class_constructor(x = make_pt_ready4class_constructor())

make_pt_ready4class_constructor(
  make_s3_lgl = logical(0),
  name_stub_chr = character(0),
  pt_ls = list(),
  pt_chkr_pfx_ls = list(),
  pt_ns_ls = list(),
  vals_ls = list(),
  allowed_vals_ls = list(),
  min_max_vals_ls = list(),
  start_end_vals_ls = list(),
  class_desc_chr = character(0),
  parent_class_chr = character(0),
  slots_ls = list(),
  meaningful_nms_ls = list(),
  inc_clss_ls = list(),
  asserts_ls = list()
)

is_ready4class_constructor(x)
}
\arguments{
\item{x}{An object of any type}

\item{make_s3_lgl}{Make S3 (a logical vector), Default: logical(0)}

\item{name_stub_chr}{Name stub (a character vector), Default: character(0)}

\item{pt_ls}{Prototype (a list), Default: list()}

\item{pt_chkr_pfx_ls}{Prototype checker prefix (a list), Default: list()}

\item{pt_ns_ls}{Prototype namespace (a list), Default: list()}

\item{vals_ls}{Values (a list), Default: list()}

\item{allowed_vals_ls}{Allowed values (a list), Default: list()}

\item{min_max_vals_ls}{Minimum maximum values (a list), Default: list()}

\item{start_end_vals_ls}{Start end values (a list), Default: list()}

\item{class_desc_chr}{Class description (a character vector), Default: character(0)}

\item{parent_class_chr}{Parent class (a character vector), Default: character(0)}

\item{slots_ls}{Slots (a list), Default: list()}

\item{meaningful_nms_ls}{Meaningful names (a list), Default: list()}

\item{inc_clss_ls}{Include classes (a list), Default: list()}

\item{asserts_ls}{Asserts (a list), Default: list()}
}
\value{
A validated instance of the Class constructor table

A prototype for Class constructor table

A logical value, TRUE if a valid instance of the Class constructor table
}
\description{
Create a new valid instance of the Class constructor table
}
\details{
Class constructor table
}
