% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_manifest.R
\name{ready4class_manifest}
\alias{ready4class_manifest}
\title{ready4 S3 class Manifest.}
\usage{
ready4class_manifest(x = make_pt_ready4class_manifest())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class Manifest., Default: make_pt_ready4class_manifest()}
}
\value{
A validated instance of the ready4 S3 class Manifest.
}
\description{
Create a new valid instance of the ready4 S3 class Manifest.
}
\details{
ready4 S3 class Manifest.
}
