% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_remake_ls_cols.R
\name{remake_ls_cols}
\alias{remake_ls_cols}
\alias{remake_ls_cols.ready4class_constructor_tbl}
\alias{remake_ls_cols,ready4class_constructor_tbl-method}
\title{Remake list columns}
\usage{
remake_ls_cols(x, ...)

remake_ls_cols.ready4class_constructor_tbl(x)

\S4method{remake_ls_cols}{ready4class_constructor_tbl}(x)
}
\arguments{
\item{x}{An instance of ready4 S3 class Constructor Table of metadata required to make new classes.}

\item{...}{Additional arguments (an additional arguments)}
}
\value{
Instance (ready4 S3 class Constructor Table of metadata required to make new classes.)
}
\description{
remake_ls_cols() is a Remake List Columns generic that remakes list columns.

remake_ls_cols.ready4class_constructor_tbl() is a Remake List Columns method that remakes list columns. This method is implemented for the ready4 S3 class Constructor Table of metadata required to make new classes.. The function is called for its side effects and does not return a value.
}
