% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_remake_ls_cols.R
\name{remake_ls_cols.ready4_constructor_tbl}
\alias{remake_ls_cols.ready4_constructor_tbl}
\title{Remake list columns method applied to readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE.}
\usage{
\method{remake_ls_cols}{ready4_constructor_tbl}(x)
}
\arguments{
\item{x}{An instance of readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE}
}
\value{
Instance (a readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE)
}
\description{
remake_ls_cols.ready4_constructor_tbl() is a Remake List Columns method that remakes list columns. This method is implemented for the readyforwhatsnext S3 class CLASS CONSTRUCTOR TABLE.NA
}
