% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_class_from_cache}
\alias{remove_class_from_cache}
\title{Remove class from cache}
\usage{
remove_class_from_cache(class_name_chr, class_in_cache_logic_chr = "stop")
}
\arguments{
\item{class_name_chr}{Class name (a character vector of length 1)}

\item{class_in_cache_logic_chr}{Class in cache logic (a character vector of length 1), Default: 'stop'}
}
\description{
remove_class_from_cache() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove class from cache. Function argument class_name_chr specifies the object to be updated. Argument class_in_cache_logic_chr provides the object to be updated.NA
}
\keyword{internal}
