% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew.ready4class_constructor}
\alias{renew.ready4class_constructor}
\alias{renew,ready4class_constructor-method}
\alias{renew.ready4class_pt_lup}
\alias{renew,ready4class_pt_lup-method}
\title{Renew method applied to ready4 S3 class Constructor Table..}
\usage{
renew.ready4class_constructor(
  x,
  type_1L_chr = "listify",
  name_pfx_1L_chr = NULL
)

\S4method{renew}{ready4class_constructor}(x, type_1L_chr = "listify", name_pfx_1L_chr = NULL)

renew.ready4class_pt_lup(x, attached_nss_chr, type_1L_chr = "namespace")

\S4method{renew}{ready4class_pt_lup}(x, attached_nss_chr, type_1L_chr = "namespace")
}
\arguments{
\item{x}{An instance of ready4 S3 class Prototype Lookup Table of class metadata.}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'namespace'}

\item{name_pfx_1L_chr}{Name prefix (a character vector of length one), Default: NULL}

\item{attached_nss_chr}{Attached namespaces (a character vector)}
}
\value{
NA ()

NA ()
}
\description{
renew.ready4class_constructor() is a Renew method that updates an instance of a class with new values. This method is implemented for the ready4 S3 class Constructor Table. The function is called for its side effects and does not return a value.

renew.ready4class_pt_lup() is a Renew method that updates an instance of a class with new values. This method is implemented for the ready4 S3 class Prototype Lookup Table of class metadata. The function is called for its side effects and does not return a value.
}
