% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_replace.R
\name{replace_NA_in_fn}
\alias{replace_NA_in_fn}
\title{Replace NA in}
\usage{
replace_NA_in_fn(fn_chr)
}
\arguments{
\item{fn_chr}{Function (a character vector)}
}
\description{
replace_NA_in_fn() is a Replace function that edits an object, replacing a specified element with another specified element. Specifically, this function implements an algorithm to a replace NA in. Function argument fn_chr specifies the object to be updated. Argument NA provides the object to be updated.The function is called for its side effects and does not return a value.
}
