% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_alg_to_ref_cls_nm}
\alias{transform_alg_to_ref_cls_nm}
\title{Transform algorithm to reference class name}
\usage{
transform_alg_to_ref_cls_nm(alg_to_ref_cls_nm, pkg_nm_1L_chr)
}
\arguments{
\item{alg_to_ref_cls_nm}{PARAM_DESCRIPTION}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one)}
}
\description{
transform_alg_to_ref_cls_nm() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform algorithm to reference class name. Function argument alg_to_ref_cls_nm specifies the object to be updated. Argument pkg_nm_1L_chr provides the object to be updated. The function is called for its side effects and does not return a value.
}
