% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_pt_ls_for_new_clss}
\alias{transform_pt_ls_for_new_clss}
\title{Transform prototype list for new classes}
\usage{
transform_pt_ls_for_new_clss(pts_for_new_clss_ls)
}
\arguments{
\item{pts_for_new_clss_ls}{Prototypes for new classes (a list)}
}
\value{
Tfd prototypes for new clss (a list)
}
\description{
transform_pt_ls_for_new_clss() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to a transform prototype list for new clss. Function argument pts_for_new_clss_ls specifies the object to be updated. Argument NA provides the object to be updated.The function returns tfd prototypes for new clss (a list).
}
\keyword{internal}
