% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_update_lup_for_ns.R
\name{update_lup_for_ns}
\alias{update_lup_for_ns}
\alias{update_lup_for_ns.ready4_class_pt_lup}
\alias{update_lup_for_ns,ready4_class_pt_lup-method}
\title{Update lookup table for namespace}
\usage{
update_lup_for_ns(x, attached_nss_chr, ...)

update_lup_for_ns.ready4_class_pt_lup(x, attached_nss_chr)

\S4method{update_lup_for_ns}{ready4_class_pt_lup}(x, attached_nss_chr)
}
\arguments{
\item{x}{An instance of ready4 S3 class CLASS PROTOTYPE LOOKUP TABLE}

\item{attached_nss_chr}{Attached namespaces (a character vector)}

\item{...}{Additional arguments}
}
\value{
Instance (a ready4 S3 class CLASS PROTOTYPE LOOKUP TABLE)
}
\description{
update_lup_for_ns() is an Update Lookup Table for Namespace generic that updates a lookup table with namespace data.

update_lup_for_ns.ready4_class_pt_lup() is an Update Lookup Table for Namespace method that updates a lookup table with namespace data. This method is implemented for the ready4 S3 class CLASS PROTOTYPE LOOKUP TABLE. The function returns Instance (a ready4 S3 class CLASS PROTOTYPE LOOKUP TABLE).
}
