% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_update_lup_for_ns.R
\name{update_lup_for_ns.ready4_class_pt_lup}
\alias{update_lup_for_ns.ready4_class_pt_lup}
\title{Update lookup table for namespace method applied to readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE.}
\usage{
update_lup_for_ns.ready4_class_pt_lup(x, attached_nss_chr)
}
\arguments{
\item{x}{An instance of readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE}

\item{attached_nss_chr}{Attached namespaces (a character vector)}
}
\value{
Instance (a readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE)
}
\description{
update_lup_for_ns.ready4_class_pt_lup() is an Update Lookup Table for Namespace method that updates a lookup table with namespace data. This method is implemented for the readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE. The function returns Instance (a readyforwhatsnext S3 class CLASS PROTOTYPE LOOKUP TABLE).
}
