% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_class_pt_lup.R
\name{validate_ready4_class_pt_lup}
\alias{validate_ready4_class_pt_lup}
\title{Validate CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class}
\usage{
validate_ready4_class_pt_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class}
}
\value{
A prototpe for CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class
}
\description{
Validate an instance of the CLASS PROTOTYPE LOOKUP TABLE readyforwhatsnext S3 class
}
\details{
CLASS PROTOTYPE LOOKUP TABLE Readyforwhatsnext S3 class of metadata of prototype classes.
}
