% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_constructor.R
\name{validate_ready4class_constructor}
\alias{validate_ready4class_constructor}
\title{Validate ready4class package constructor ready4 S3 class Constructor Table.}
\usage{
validate_ready4class_constructor(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class Constructor Table.}
}
\value{
A prototpe for ready4 S3 class Constructor Table.
}
\description{
Validate an instance of the ready4 S3 class Constructor Table.
}
\details{
ready4 S3 class Constructor Table.
}
