% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_manifest.R
\name{validate_ready4class_manifest}
\alias{validate_ready4class_manifest}
\title{Validate ready4class package manifest ready4 S3 class Manifest.}
\usage{
validate_ready4class_manifest(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class Manifest.}
}
\value{
A prototpe for ready4 S3 class Manifest.
}
\description{
Validate an instance of the ready4 S3 class Manifest.
}
\details{
ready4 S3 class Manifest.
}
