% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4class_pt_lup.R
\name{validate_ready4class_pt_lup}
\alias{validate_ready4class_pt_lup}
\title{validate ready4class package prototype lookup table ready4 S3 class Prototype Lookup Table of class metadata.}
\usage{
validate_ready4class_pt_lup(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class Prototype Lookup Table of class metadata.}
}
\value{
A prototpe for ready4 S3 class Prototype Lookup Table of class metadata.
}
\description{
Validate an instance of the ready4 S3 class Prototype Lookup Table of class metadata.
}
\details{
ready4 S3 class Prototype Lookup Table.
}
\keyword{internal}
