% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{write_classes}
\alias{write_classes}
\title{Write classes}
\usage{
write_classes(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\description{
write_classes() is a Write Classes generic that writes new classes.NA
}
