% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_gtr_str_mthds_for_r4}
\alias{write_gtr_str_mthds_for_r4}
\title{Write getter setter methods for}
\usage{
write_gtr_str_mthds_for_r4(
  slot_name_chr,
  set_only,
  import_packages_ls,
  class_name,
  print_accessors,
  output_folder
)
}
\arguments{
\item{slot_name_chr}{Slot name (a character vector)}

\item{set_only}{PARAM_DESCRIPTION}

\item{import_packages_ls}{Import packages (a list)}

\item{class_name}{PARAM_DESCRIPTION}

\item{print_accessors}{PARAM_DESCRIPTION}

\item{output_folder}{PARAM_DESCRIPTION}
}
\description{
write_gtr_str_mthds_for_r4() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write a getter setter mthds for.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
