% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_script_to_make_gnrc}
\alias{write_script_to_make_gnrc}
\title{Write script to make generic}
\usage{
write_script_to_make_gnrc(
  write_file_ls,
  generic_exists_lgl,
  gen_mthd_pair_ls,
  fn_name_chr,
  fn_type_chr,
  fn_desc_chr = NA_character_,
  fn_out_type_chr = NA_character_,
  fn_title_chr = NA_character_,
  class_name_chr = NA_character_,
  output_dir_chr = NA_character_,
  overwrite_lgl = F,
  s3_lgl = F,
  write_lgl = T,
  doc_in_class_lgl = F
)
}
\arguments{
\item{write_file_ls}{Write file (a list)}

\item{generic_exists_lgl}{Generic exists (a logical vector)}

\item{gen_mthd_pair_ls}{Generate method pair (a list)}

\item{fn_name_chr}{Function name (a character vector)}

\item{fn_type_chr}{Function type (a character vector)}

\item{fn_desc_chr}{Function description (a character vector), Default: 'NA'}

\item{fn_out_type_chr}{Function out type (a character vector), Default: 'NA'}

\item{fn_title_chr}{Function title (a character vector), Default: 'NA'}

\item{class_name_chr}{Class name (a character vector), Default: 'NA'}

\item{output_dir_chr}{Output directory (a character vector), Default: 'NA'}

\item{overwrite_lgl}{Overwrite (a logical vector), Default: F}

\item{s3_lgl}{S3 (a logical vector), Default: F}

\item{write_lgl}{Write (a logical vector), Default: T}

\item{doc_in_class_lgl}{Document in class (a logical vector), Default: F}
}
\description{
write_script_to_make_gnrc() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write script to a make gnrc.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
