% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_scripts_to_mk_clss}
\alias{write_scripts_to_mk_clss}
\title{write scripts to make classes}
\usage{
write_scripts_to_mk_clss(
  pts_for_new_clss_ls,
  pkg_nm_1L_chr,
  class_pfx_1L_chr,
  R_dir_1L_chr = "R",
  pt_lup,
  description_ls = NULL,
  nss_to_ignore_chr = NA_character_,
  req_pkgs_chr = NA_character_
)
}
\arguments{
\item{pts_for_new_clss_ls}{Prototypes for new classes (a list)}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one)}

\item{class_pfx_1L_chr}{Class prefix (a character vector of length one)}

\item{R_dir_1L_chr}{R directory (a character vector of length one), Default: 'R'}

\item{pt_lup}{Prototype (a lookup table)}

\item{description_ls}{Description (a list), Default: NULL}

\item{nss_to_ignore_chr}{Namespaces to ignore (a character vector), Default: 'NA'}

\item{req_pkgs_chr}{Require packages (a character vector), Default: 'NA'}
}
\description{
write_scripts_to_mk_clss() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write scripts to make classes. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
\keyword{internal}
