% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_std_mthd}
\alias{write_std_mthd}
\title{Write standard method}
\usage{
write_std_mthd(
  fn,
  fn_name_1L_chr,
  class_nm_1L_chr,
  fn_desc_chr,
  fn_title_1L_chr,
  fn_outp_type_1L_chr,
  fn_types_lup,
  pkg_nm_1L_chr,
  output_dir_1L_chr,
  signature_1L_chr = NA_character_,
  append_1L_lgl = T,
  first_1L_lgl = T,
  import_from_chr = NA_character_
)
}
\arguments{
\item{fn}{Function (a function)}

\item{fn_name_1L_chr}{Function name (a character vector of length one)}

\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{fn_desc_chr}{Function description (a character vector)}

\item{fn_title_1L_chr}{Function title (a character vector of length one)}

\item{fn_outp_type_1L_chr}{Function output type (a character vector of length one)}

\item{fn_types_lup}{Function types (a lookup table)}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one)}

\item{output_dir_1L_chr}{Output directory (a character vector of length one)}

\item{signature_1L_chr}{Signature (a character vector of length one), Default: 'NA'}

\item{append_1L_lgl}{Append (a logical vector of length one), Default: T}

\item{first_1L_lgl}{First (a logical vector of length one), Default: T}

\item{import_from_chr}{Import from (a character vector), Default: 'NA'}
}
\value{
No return value, called for side effects.
}
\description{
write_std_mthd() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write standard method. The function is called for its side effects and does not return a value.
}
\keyword{internal}
