% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_delete_gnrc_fn_fls}
\alias{write_to_delete_gnrc_fn_fls}
\title{Write to delete generic function files}
\usage{
write_to_delete_gnrc_fn_fls(x, output_dir_1L_chr)
}
\arguments{
\item{x}{An object}

\item{output_dir_1L_chr}{Output directory (a character vector of length one)}
}
\value{
No return value, called for side effects.
}
\description{
write_to_delete_gnrc_fn_fls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to delete generic function files. The function is called for its side effects and does not return a value.
}
\keyword{internal}
