% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_mk_r4_cls}
\alias{write_to_mk_r4_cls}
\title{Write to make readyforwhatsnext S4 class}
\usage{
write_to_mk_r4_cls(
  class_nm_1L_chr,
  slots_chr,
  type_chr,
  proto_ls,
  parent_cls_nm_1L_chr,
  print_set_cls_1L_lgl,
  class_desc_1L_chr,
  output_file_class,
  clss_to_inc_chr,
  prototype_lup,
  helper_lgl = F,
  parent_ns_ls
)
}
\arguments{
\item{class_nm_1L_chr}{Class name (a character vector of length one)}

\item{slots_chr}{Slots (a character vector)}

\item{type_chr}{Type (a character vector)}

\item{proto_ls}{Proto (a list)}

\item{parent_cls_nm_1L_chr}{Parent class name (a character vector of length one)}

\item{print_set_cls_1L_lgl}{Print set class (a logical vector of length one)}

\item{class_desc_1L_chr}{Class description (a character vector of length one)}

\item{output_file_class}{PARAM_DESCRIPTION}

\item{clss_to_inc_chr}{Classes to inc (a character vector)}

\item{prototype_lup}{Prototype (a lookup table)}

\item{helper_lgl}{Helper (a logical vector), Default: F}

\item{parent_ns_ls}{Parent namespace (a list)}
}
\description{
write_to_mk_r4_cls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to make readyforwhatsnext s4 class.The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
